Subset of daily COVID-19 doctor visits and cases from 6 states in archive format

This data source is based on information about outpatient visits, provided to us by health system partners, and also contains confirmed COVID-19 cases based on reports made available by the Center for Systems Science and Engineering at Johns Hopkins University. This example data ranges from June 1, 2020 to December 1, 2021, issued on dates from June 1, 2020 to December 1, 2021. It is limited to California, Florida, Texas, and New York.

It is used in the epiprocess compactify, epi_archive, and advanced-use (advanced) vignettes.

Usage

archive_cases_dv_subset

Format

An object of class epi_archive of length 6.

Source

This object contains a modified part of the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University as republished in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by Johns Hopkins University on behalf of its Center for Systems Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:

From the COVIDcast Epidata API: case_rate_7d_av signal was computed by Delphi from the original JHU-CSSE data by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.

Furthermore, the data has been limited to a very small number of rows, the signal names slightly altered, and formatted into an epi_archive.

This object contains a modified part of the Delphi doctor-visits indicator. This data source is computed by the Delphi Group from information about outpatient visits, provided to Delphi by health system partners, and published in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by the Delphi group.

Modifications:

From the COVIDcast Doctor Visits signal: The signal smoothed_adj_cli is taken directly from the API without changes.

Furthermore, the data has been limited to a very small number of rows, the signal names slightly altered, and formatted into an epi_archive.

Data dictionary

The data in the epi_archive$DT attribute has columns:

geo_value

the geographic value associated with each row of measurements.

time_value

the time value associated with each row of measurements.

version

the time value specifying the version for each row of measurements.

percent_cli

percentage of doctor’s visits with CLI (COVID-like illness) computed from medical insurance claims

case_rate_7d_av

7-day average signal of number of new confirmed cases due to COVID-19 per 100,000 population, daily

Examples

# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epiprocess::archive_cases_dv_subset
#> → An `epi_archive` object, with metadata:
#> ℹ Min/max time values: 2020-06-01 / 2021-11-30
#> ℹ First/last version with update: 2020-06-02 / 2021-12-01
#> ℹ Versions end: 2021-12-01
#> ℹ A preview of the table (129638 rows x 5 columns):
#> Key: <geo_value, time_value, version>
#>         geo_value time_value    version percent_cli case_rate_7d_av
#>            <char>     <Date>     <Date>       <num>           <num>
#>      1:        ca 2020-06-01 2020-06-02          NA        6.628329
#>      2:        ca 2020-06-01 2020-06-06    2.140116        6.628329
#>      3:        ca 2020-06-01 2020-06-07    2.140116        6.628329
#>      4:        ca 2020-06-01 2020-06-08    2.140379        6.628329
#>      5:        ca 2020-06-01 2020-06-09    2.114430        6.628329
#>     ---                                                            
#> 129634:        tx 2021-11-26 2021-11-29    1.858596        7.957657
#> 129635:        tx 2021-11-27 2021-11-28          NA        7.174299
#> 129636:        tx 2021-11-28 2021-11-29          NA        6.834681
#> 129637:        tx 2021-11-29 2021-11-30          NA        8.841247
#> 129638:        tx 2021-11-30 2021-12-01          NA        9.566218

# works
library(epiprocess)
archive_cases_dv_subset
#> → An `epi_archive` object, with metadata:
#> ℹ Min/max time values: 2020-06-01 / 2021-11-30
#> ℹ First/last version with update: 2020-06-02 / 2021-12-01
#> ℹ Versions end: 2021-12-01
#> ℹ A preview of the table (129638 rows x 5 columns):
#> Key: <geo_value, time_value, version>
#>         geo_value time_value    version percent_cli case_rate_7d_av
#>            <char>     <Date>     <Date>       <num>           <num>
#>      1:        ca 2020-06-01 2020-06-02          NA        6.628329
#>      2:        ca 2020-06-01 2020-06-06    2.140116        6.628329
#>      3:        ca 2020-06-01 2020-06-07    2.140116        6.628329
#>      4:        ca 2020-06-01 2020-06-08    2.140379        6.628329
#>      5:        ca 2020-06-01 2020-06-09    2.114430        6.628329
#>     ---                                                            
#> 129634:        tx 2021-11-26 2021-11-29    1.858596        7.957657
#> 129635:        tx 2021-11-27 2021-11-28          NA        7.174299
#> 129636:        tx 2021-11-28 2021-11-29          NA        6.834681
#> 129637:        tx 2021-11-29 2021-11-30          NA        8.841247
#> 129638:        tx 2021-11-30 2021-12-01          NA        9.566218

# fails
if (FALSE) { # \dontrun{
data(archive_cases_dv_subset, package = "epiprocess")
} # }