Convert data frame to epi_signal format

Converts a data frame or tibble into a format consistent with the epi_signal class (ensuring that it has a certain minimal set of columns, and that it has certain metadata stored in its attributes).

as.epi_signal(x, ...)

# S3 method for epi_signal
as.epi_signal(x, ...)

# S3 method for tibble
as.epi_signal(
  x,
  name,
  geo_type,
  time_type,
  signal_type,
  signal_unit,
  issue,
  metadata = list(),
  ...
)

# S3 method for data.frame
as.epi_signal(
  x,
  name,
  geo_type,
  time_type,
  signal_type,
  signal_unit,
  issue,
  metadata = list(),
  ...
)

Arguments

x
Object to be converted. See Methods section below for details on formatting of each input type.

...
Additional arguments passed to methods.

name
The name to use for this signal.

geo_type
The geographic resolution. If missing, then it will be guessed from the geo values present.

time_type
The temporal resolution. If missing, then it will be guessed from the time values present.

signal_type
The type of the signal value.

signal_unit
The units of the signal value.

issue
Issue date to use for this data, if not present in x. If no issue date is present in x and issue is missing, then today's date will be used.

metadata
List or tibble of additional metadata to attach to the epi_signal object. All objects will have geo_type, time_type, signal_type (optional), and signal_unit (optional) entries included in their metadata, derived from the above arguments; any entries in the passed argument will be preserved in the metadata as well.

Value

An epi_signal object.

Details

An epi_signal object is simply a tibble, with (at least) the following columns (with data types written in tibble notation):

value <dbl>: the value of the signal

geo_value <int> or <str>: the associated geographic value

time_value <date>: the associated time value

issue <date>: the time value at which the given signal value was issued

An epi_signal object also has a tibble metadata stored in its attributes, with (at least) the following columns:

name <str>: the name of the signal

geo_type <str>: the geographic resolution

time_type <str>: the temporal resolution

signal_type <str>: the type of the signal value (optional)

signal_unit <str>: the units associated with the signal value (optional)

More information on geo types, time types, and signal types, is given below.

Methods (by class)

epi_signal: Simply returns the epi_signal object unchanged.

tibble: The input tibble x must contain the columns value, geo_value, and time_value. If an issue column is present in x, it will be used as the issue date for each observation; if not, the issue argument will be used. Other columns will be preserved as-is.

data.frame: The input data frame x must contain the columns value, geo_value, and time_value. If an issue column is present in x, it will be used as the issue date for each observation; if not, the issue argument will be used. Other columns will be preserved as-is.

Geo types

The allowable geo types, and their coding (allowable range for geo values), are as follows.

county: U.S. counties; coded by 5-digit FIPS codes.

hrr: U.S. hospital referral regions (designed to represent regional health care markets); there are 306 HRRs in the U.S; coded by number (nonconsecutive, between 1 and 457).

state: U.S. states; coded by 2-digit postal abbreviation (lowercase); note that Puerto Rico is "pr" and Washington D.C. is "dc".

hhs: U.S. HHS regions; coded by number (between 1 and 10).

nation: country; coded by ISO 3166-1 alpha-2 country codes (lowercase).

Time types

The allowable time types are as follows. In each case, their coding (allowable range for time values) is an object of class Date, or string in the format "YYYY-MM-DD".

"day": each observation covers one day.

"week": each observation covers one week, and the time value is assumed to be the start date of the epiweek (MMWR week) that the data represents.

Signal types

todo