Convert data to epi_tibble format

Converts a data frame or tibble into a format that is consistent with the epi_tibble class, ensuring that it has a certain minimal set of columns, and that it has certain minimal metadata.

as.epi_tibble(x, ...)

# S3 method for epi_tibble
as.epi_tibble(x, ...)

# S3 method for tibble
as.epi_tibble(x, geo_type, time_type, issue, additional_metadata = list(), ...)

# S3 method for data.frame
as.epi_tibble(x, geo_type, time_type, issue, additional_metadata = list(), ...)

Arguments

x
The object to be converted. See the methods section below for details on formatting of each input type.

...
Additional arguments passed to methods.

geo_type
The type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type
The type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

issue
Issue to use for this data. If missing, then the function will attempt to infer it from the passed object x; if this fails, then the current day-time will be used.

additional_metadata
List of additional metadata to attach to the epi_tibble object. All objects will have time_type, geo_type, and issue fields; named entries from the passed list or will be included as well.

Value

An epi_tibble object.

Details

An epi_tibble is a tibble with (at least) the following columns:

geo_value: the geographic value associated with each measurement.

time_value: the time value associated with each measurement.

Other columns can be considered as measured variables, which we also broadly refer to as signal variables. An epi_tibble object also has metadata with (at least) the following fields:

geo_type: the type for the geo values.

time_type: the type for the time values.

issue: the time value at which the given data set was issued.

The first two fields above, geo_type and time_type, can usually be inferred from the geo_value and time_value columns, respectively. The last field above, issue, is the most unique to the epi_tibble format. In a typical case, this represents the maximum of the issues of individual signal values measured in the data set; hence we would also say that the data set is comprised of all signal values observed "as of" the given issue in the metadata.

Metadata for an epi_tibble object x can be accessed (and altered) via attributes(x)$metadata. More information on geo types, time types, and issues is given below.

Methods (by class)

epi_tibble: Simply returns the epi_tibble object unchanged.

tibble: The input tibble x must contain the columns geo_value and time_value. All other columns will be preserved as is, and treated as measured variables. If issue is missing, then the function will look for issue as a column of x, or as a field in its metadata (stored in its attributes), to infer the issue; if this fails, then the current day-time will be used.

data.frame: The input data frame x must contain the columns geo_value and time_value. All other columns will be preserved as is, and treated as measured variables. If issue is missing, then the function will look for issue as a column of x, or as a field in its metadata (stored in its attributes), to infer the issue; if this fails, then the current day-time will be used.

Geo types

The following geo types are supported in an epi_tibble. Their geo coding (specification of geo values for each geo type) is also described below.

"county": each observation corresponds to a U.S. county; coded by 5-digit FIPS code.

"hrr": each observation corresponds to a U.S. hospital referral region (designed to represent regional healthcare markets); there are 306 HRRs in the U.S; coded by number (nonconsecutive, between 1 and 457).

"state": each observation corresponds to a U.S. state; coded by 2-digit postal abbreviation (lowercase); note that Puerto Rico is "pr" and Washington D.C. is "dc".

"hhs": each observation corresponds to a U.S. HHS region; coded by number (consecutive, between 1 and 10).

"nation": each observation corresponds to a country; coded by ISO 31661- alpha-2 country codes (lowercase).

The above geo types come with aggregation utilities in the package, todo: refer to relevant functionality, vignette, and so on. An unrecognizable geo type is labeled as "custom".

Time types

The following time types are supported in an epi_tibble. Their time coding (specification of time values for each time type) is also described below.

"day-time": each observation corresponds to a time on a given day (measured to the second); coded as a POSIXct object, as in as.POSIXct("2020-06-09 18:45:40").

"day": each observation corresponds to a day; coded as a Date object, as in as.Date("2020-06-09").

"week": each observation corresponds to a week; the alignment can be arbitrary (as to whether a week starts on a Monday, Tuesday, etc.; the U.S. CDC definition of an epidemiological week starts on a Sunday); coded as a Date object, representing the start date of week.

An unrecognisable time type is labeled as "custom".

Issues

todo