Convert to epi_df format

Converts a data frame or tibble into an epi_df object. See the getting started guide for examples.

as_epi_df(x, ...)

# S3 method for epi_df
as_epi_df(x, ...)

# S3 method for tbl_df
as_epi_df(x, geo_type, time_type, as_of, additional_metadata = list(), ...)

# S3 method for data.frame
as_epi_df(x, geo_type, time_type, as_of, additional_metadata = list(), ...)

# S3 method for tbl_ts
as_epi_df(x, geo_type, time_type, as_of, additional_metadata = list(), ...)

Arguments

...

Additional arguments passed to methods.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

as_of

Time value representing the time at which the given data were available. For example, if as_of is January 31, 2022, then the epi_df object that is created would represent the most up-to-date version of the data available as of January 31, 2022. If the as_of argument is missing, then the current day-time will be used.

additional_metadata

List of additional metadata to attach to the epi_df object. The metadata will have geo_type, time_type, and as_of fields; named entries from the passed list or will be included as well.

Value

An epi_df object.

Methods (by class)

epi_df: Simply returns the epi_df object unchanged.

tbl_df: The input tibble x must contain the columns geo_value and time_value. All other columns will be preserved as is, and treated as measured variables. If as_of is missing, then the function will try to guess it from an as_of, issue, or version column of x (if any of these are present), or from as an as_of field in its metadata (stored in its attributes); if this fails, then the current day-time will be used.

data.frame: Works analogously to as_epi_df.tbl_df().

tbl_ts: Works analogously to as_epi_df.tbl_df(), except that the tbl_ts class is dropped, and any key variables (other than "geo_value") are added to the metadata of the returned object, under the other_keys field.