JHU daily COVID-19 cases and deaths rates from all states

This data source of confirmed COVID-19 cases and deaths is based on reports made available by the Center for Systems Science and Engineering at Johns Hopkins University, as downloaded from the CMU Delphi COVIDcast Epidata API. This example data is a snapshot as of May 31, 2022, and ranges from March 1, 2020 to December 31, 2021. It includes all states.

Usage

covid_case_death_rates_extended

Format

An object of class epi_df (inherits from tbl_df, tbl, data.frame) with 37576 rows and 4 columns.

Source

This object contains a modified part of the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University as republished in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by the Johns Hopkins University on behalf of its Center for Systems Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:

From the COVIDcast Epidata API: These signals are taken directly from the JHU CSSE COVID-19 GitHub repository without changes. The 7-day average signals are computed by Delphi by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.

Data dictionary

The data has columns:

geo_value

the geographic value associated with each row of measurements.

time_value

the time value associated with each row of measurements.

case_rate

7-day average signal of number of new confirmed COVID-19 cases per 100,000 population, daily

death_rate

7-day average signal of number of new confirmed deaths due to COVID-19 per 100,000 population, daily

Examples

# Since this is a re-exported dataset, it cannot be loaded using
# the `data()` function. `data()` looks for a file of the same name
# in the `data/` directory, which doesn't exist in this package.
# works
epiprocess::covid_case_death_rates_extended
#> An `epi_df` object, 37,576 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2022-05-31
#> 
#> # A tibble: 37,576 × 4
#>    geo_value time_value case_rate death_rate
#>  * <chr>     <date>         <dbl>      <dbl>
#>  1 ak        2020-03-01   0                0
#>  2 al        2020-03-01   0                0
#>  3 ar        2020-03-01   0                0
#>  4 as        2020-03-01   0                0
#>  5 az        2020-03-01   0                0
#>  6 ca        2020-03-01   0.00327          0
#>  7 co        2020-03-01   0                0
#>  8 ct        2020-03-01   0                0
#>  9 dc        2020-03-01   0                0
#> 10 de        2020-03-01   0                0
#> # ℹ 37,566 more rows

# works
library(epiprocess)
covid_case_death_rates_extended
#> An `epi_df` object, 37,576 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2022-05-31
#> 
#> # A tibble: 37,576 × 4
#>    geo_value time_value case_rate death_rate
#>  * <chr>     <date>         <dbl>      <dbl>
#>  1 ak        2020-03-01   0                0
#>  2 al        2020-03-01   0                0
#>  3 ar        2020-03-01   0                0
#>  4 as        2020-03-01   0                0
#>  5 az        2020-03-01   0                0
#>  6 ca        2020-03-01   0.00327          0
#>  7 co        2020-03-01   0                0
#>  8 ct        2020-03-01   0                0
#>  9 dc        2020-03-01   0                0
#> 10 de        2020-03-01   0                0
#> # ℹ 37,566 more rows

# fails
if (FALSE) { # \dontrun{
data(covid_case_death_rates_extended, package = "epiprocess")
} # }