epi_archive object

An epi_archive is an R6 class which contains a data table along with several relevant pieces of metadata. The data table can be seen as the full archive (version history) for some signal variables of interest.

Details

An epi_archive is an R6 class which contains a data table DT, of class data.table from the data.table package, with (at least) the following columns:

geo_value: the geographic value associated with each row of measurements.

time_value: the time value associated with each row of measurements.

version: the time value specifying the version for each row of measurements. For example, if in a given row the version is January 15, 2022 and time_value is January 14, 2022, then this row contains the measurements of the data for January 14, 2022 that were available one day later.

The data table DT has key variables geo_value, time_value, version, as well as any others (these can be specified when instantiating the epi_archive object via the other_keys argument, and/or set by operating on DT directly). There can only be a single row per unique combination of key variables, and thus the key variables are critical for figuring out how to generate a snapshot of data from the archive, as of a given version (also described below).

In general, last-observation-carried-forward (LOCF) is used to data in between recorded versions. Currently, deletions must be represented as revising a row to a special state (e.g., making the entries NA or including a special column that flags the data as removed and performing some kind of post-processing), and the archive is unaware of what this state is.

A word of caution: R6 objects, unlike most other objects in R, have reference semantics. A primary consequence of this is that objects are not copied when modified. You can read more about this in Hadley Wickham's Advanced R book.

Metadata

The following pieces of metadata are included as fields in an epi_archive object:

geo_type: the type for the geo values.

time_type: the type for the time values.

additional_metadata: list of additional metadata for the data archive.

Unlike an epi_df object, metadata for an epi_archive object x can be accessed (and altered) directly, as in x$geo_type or x$time_type, etc. Like an epi_df object, the geo_type and time_type fields in the metadata of an epi_archive object are not currently used by any downstream functions in the epiprocess package, and serve only as useful bits of information to convey about the data set at hand.

Generating Snapshots

An epi_archive object can be used to generate a snapshot of the data in epi_df format, which represents the most up-to-date values of the signal variables, as of the specified version. This is accomplished by calling the as_of() method for an epi_archive object x, for example:

x$as_of(as.Date("2022-01-15"))

to generate a snapshot as of January 15, 2022. More details on the as_of() method are documented below.

Sliding Computations

We can run a sliding computation over an epi_archive object, much like epi_slide() does for an epi_df object. This is accomplished by calling the slide() method for an epi_archive object, which works similarly to the way epi_slide() works for an epi_df object, but with one key difference: it is version-aware. That is, for an epi_archive object, the sliding computation at any given reference time point t is performed on data that would have been available as of t. More details on slide() are documented below.

Methods

Public methods

epi_archive$new()

epi_archive$print()

epi_archive$as_of()

epi_archive$slide()

epi_archive$clone()

Method new()

Creates a new epi_archive object.

Usage

epi_archive$new(x, geo_type, time_type, other_keys, additional_metadata)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

Returns

An epi_archive object.

Method print()

Usage

epi_archive$print()

Method as_of()

Generates a snapshot in epi_df format as of a given version.

Usage

epi_archive$as_of(max_version, min_time_value = -Inf)

Arguments

max_version

Time value specifying the max version to permit in the snapshot. That is, the snapshot will comprise the unique rows that represent the most up-to-date signal values, as of the specified max_version, whose time values are at least min_time_value.

min_time_value

Time value specifying the min time value to permit in the snapshot. Default is -Inf, which effectively means that there is no minimum considered.

Returns

An epi_df object.

Method slide()

Slides a given function over variables in an epi_archive object. Windows are always right-aligned, unlike epi_slide(). The other arguments are as in epi_slide(), and its documentation gives more details on their useage. The exception is the by argument, which used to specify the grouping upfront (whereas in an epi_df, this would be accomplished by a call to dplyr::group_by() that precedes a call to epi_df()). See the archive vignette for examples.

Usage

epi_archive$slide(
  f,
  ...,
  n = 14,
  complete = FALSE,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  time_step,
  by
)

Arguments

f

Function or formula to slide over variables in x. To "slide" means to apply a function or formula over a running window of n time steps (where one time step is typically one day or one week). If a function, f must take x, a data frame with the same column names as the original object; followed by any number of named arguments; and ending with .... If a formula, f can operate directly on columns accessed via .x$var, as in ~ mean(.x$var) to compute a mean of a column var over a sliding window of n time steps.

...

Additional arguments to pass to the function or formula specified via f. Alternatively, if f is missing, then the current argument is interpreted as an expression for tidy evaluation.

n

Number of time steps to use in the running window. For example, if n = 5, one time step is one day, and the alignment is "right", then to produce a value on November 5, we apply the given function or formula to data in between November 1 and 5. Default is 14.

complete

Should the slide function be run over complete windows only? Default is FALSE, which allows for computation on partial windows.

new_col_name

String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

as_list_col

Should the new column be stored as a list column? Default is FALSE, in which case a list object returned by f would be unnested (using tidyr::unnest()), and the names of the resulting columns are given by prepending new_col_name to the names of the list elements.

time_step

Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a positive integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

by

The variable(s) to group by before slide computation. If missing, then the keys in the underlying data table, excluding time_value and version, will be used for grouping.

Returns

A tibble with the grouping variables, time_value, and a new column named according to the new_col_name argument, with the slide values.

Method clone()

The objects of this class are cloneable with this method.

Usage

epi_archive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.