Detect outliers based on an STL decomposition

Detects outliers based on a seasonal-trend decomposition using LOESS (STL).

epi_detect_outlr_stl(
  x,
  var,
  n_trend = 21,
  n_seasonal = 21,
  n_threshold = 21,
  seasonal_period = NULL,
  log_transform = FALSE,
  detect_negatives = FALSE,
  detection_multiplier = 2,
  min_radius = 0,
  replacement_multiplier = 0
)

Arguments

x

The epi_df object under consideration.

var

The variable in x on which to run outlier detection.

n_trend

Number of time steps to use in the rolling window for trend. Default is 21.

n_seasonal

Number of time steps to use in the rolling window for seasonality. Default is 21.

n_threshold

Number of time steps to use in rolling window for the IQR outlier thresholds.

seasonal_period

Integer specifying period of seasonality. For example, for daily data, a period 7 means weekly seasonality. The default is NULL, meaning that no seasonal term will be included in the STL decomposition.

Value

A tibble with the same number of rows as x, and columns lower, upper, and replacement.

Details

The STL decomposition is computed using the feasts package. Once computed, the outlier detection method is analogous to the rolling median method in epi_detect_outlr_rm(), except with the fitted values and residuals from the STL decomposition taking the place of the rolling median and residuals to the rolling median, respectively.

The last set of arguments, log_transform through replacement_multiplier, are exactly as in epi_detect_outlr_rm(); refer to its help file for their description.