epiprocess: Tools for basic signal processing in epidemiology

This package introduces a common data structure for epidemiological data sets measured over space and time, and offers associated utilities to perform basic signal processing tasks.

See also

Author

Maintainer: Logan Brooks lcbrooks@andrew.cmu.edu

Authors:

Daniel McDonald

Evan Ray

Ryan Tibshirani

Other contributors:

Jacob Bien [contributor]

Rafael Catoia [contributor]

Nat DeFries [contributor]

Rachel Lobay [contributor]

Ken Mawer [contributor]

Chloe You [contributor]

Quang Nguyen [contributor]

Dmitry Shemetov [contributor]

Lionel Henry (Author of included rlang fragments) [contributor]

Hadley Wickham (Author of included rlang fragments) [contributor]

Posit (Copyright holder of included rlang fragments) [copyright holder]