epiprocess: Tools for basic signal processing in epidemiology

This package introduces common data structures for working with epidemiological data reported by location and time and offers associated utilities to perform basic signal processing tasks. The package is designed to be used in conjunction with `epipredict` for building and evaluating epidemiological models.

See also

Author

Maintainer: Logan Brooks lcbrooks@andrew.cmu.edu

Authors:

Daniel McDonald

Evan Ray

Ryan Tibshirani

Other contributors:

Jacob Bien [contributor]

Rafael Catoia [contributor]

Nat DeFries [contributor]

Rachel Lobay [contributor]

Ken Mawer [contributor]

Chloe You [contributor]

Quang Nguyen [contributor]

Dmitry Shemetov [contributor]

Lionel Henry (Author of included rlang fragments) [contributor]

Hadley Wickham (Author of included rlang fragments) [contributor]

Posit (Copyright holder of included rlang fragments) [copyright holder]

Johns Hopkins University Center for Systems Science and Engineering (Owner of COVID-19 cases and deaths data from the COVID-19 Data Repository) [data contributor]

Johns Hopkins University (Copyright holder of COVID-19 cases and deaths data from the COVID-19 Data Repository) [copyright holder]

Carnegie Mellon University Delphi Group (Owner of claims-based CLI data from the Delphi Epidata API) [data contributor]