Generate a snapshot from an epi_archive object

Generates a snapshot in epi_df format from an epi_archive object, as of a given version. See the archive vignette for examples.

epix_as_of(x, max_version, min_time_value = -Inf)

Arguments

max_version
Time value specifying the max version to permit in the snapshot. That is, the snapshot will comprise the unique rows of the current archive data that represent the most up-to-date signal values, as of the specified max_version (and whose time values are at least min_time_value.)

min_time_value
Time value specifying the min time value to permit in the snapshot. Default is -Inf, which effectively means that there is no minimum considered.

Value

An epi_df object.

Details

This is simply a wrapper around the as_of() method of the epi_archive class, so if x is an epi_archive object, then:

epix_as_of(x, max_version = v)

is equivalent to:

x$as_of(max_version = v)