Estimate derivative

Estimates the derivative at the end of a sequence of values, using linear regression, a smoothing spline, or trend filtering. This is a convenience function to be used inside of a call to epi_slide(), see the estimating derivatives vignette for examples.

estimate_deriv(
  var,
  method = c("lin", "ss", "tf"),
  deriv = 1,
  na_rm = FALSE,
  keep_obj = FALSE,
  ...
)

Arguments

var

The variable whose derivative is to be estimated.

method

One of "lin", "ss", or "tf" indicating the method to use for the derivative calculation. We run the given method using var as the response variable, assuming an equally-spaced sequence of design points, and return the corresponding estimated derivative (that is, the derivative of the underlying estimated function, linear or spline) at the end of the sequence. Default is "lin". See details for more explanation.

deriv

Order of derivative to estimate. Only orders 1 or 2 are supported, with the default being 1. (In some cases, a second-order derivative will return a trivial result; for example: when method is "lin", this will always be zero.)

na_rm

Should missing values be removed before the computation? Default is FALSE.

keep_obj

Should the fitted object (from linear regression, smoothing spline, or trend filtering) be returned, along with the estimated derivative, in a list? This is typically to perform some kind of post hoc inspection. Default is FALSE.

...

Additional arguments to pass to the method used to estimate the derivative. See details below.

Details

The derivative is estimated using:

Linear regression, when method is "lin", via stats::lsfit().

Cubic smoothing spline, when method is "ss", via stats::smooth.spline().

Polynomial trend filtering, when method is "tf", via genlasso::trendfilter().

The second and third cases base the derivative calculation on a nonparametric fit and should typically be used with a larger window n in a parent call to epi_slide(). The third case (trend filtering) is more locally adaptive than the second (smoothing spline) and can work better when there are sharp changes in the smoothness of the underlying values.

In the first and second cases (linear regression and smoothing spline), the additional arguments in ... are directly passed to the underlying estimation function (stats::lsfit() and stats::smooth.spline()).

The third case (trend filtering) works a little differently: here, a custom set of arguments is allowed (and are internally distributed as appropriate to the functions genlasso::trendfilter(), genlasso::cv.trendfilter(), and genlasso::coef.genlasso()):

ord

Order of piecewise polynomial for the trend filtering fit, default is 2.

maxsteps

Maximum number of steps to take in the solution path before terminating, default is 100.

cv

Boolean indicating whether cross-validation should be used to choose an effective degrees of freedom for the fit, default is FALSE.

k

Number of folds if cross-validation is to be used. Default is 5.

df

Desired effective degrees of freedom for the trend filtering fit. If cv = FALSE, then df must be an integer; if cv = TRUE, then df should be one of "min" or "1se" indicating the selection rule to use based on the cross-validation error curve (minimum or 1-standard-error rule, respectively). Default is 8 when cv = FALSE, and "1se" when cv = TRUE.