Run FASTQC

Run FASTQC on fastq files.

fastqc(
  files,
  rootdir = "./",
  outdir = file.path(rootdir, "processed_data", "batch1", "fastqFileQC", basename(fq)),
  dir = "./fastqc_temp",
  format = "fastq",
  threads = 1,
  fastqc_x = echoconda::find_packages("fastqc", conda_env = "epiprocess")$path,
  ...
)

Arguments

files

Paths to files of the type specified in format.

rootdir

Base directory containing all files for the project.

outdir

Create all output files in the specified output directory. Please note that this directory must exist as the program will not create it. If this option is not set then the output file for each sequence file is created in the same directory as the sequence file which was processed.

dir

Selects a directory to be used for temporary files written when generating report images. Defaults to system temp directory if not specified.

format

Bypasses the normal sequence file format detection and forces the program to use the specified format. Valid formats are bam,sam,bam_mapped,sam_mapped and fastq.

threads

Specifies the number of files which can be processed simultaneously. Each thread will be allocated 250MB of memory so you shouldn't run more threads than your available memory will cope with, and not more than 6 threads on a 32 bit machine.

fastqc_x

Path to fastqc software executable.

...

Additional string arguments to be passed to fastqc.