Compute correlations between variables in an epi_tibble object

Computes correlations between variables in an epi_tibble object, allowing for grouping by geo value, time value, other variables. See the correlations vignette for examples.

grouped_cor(
  x,
  var1,
  var2,
  dt1 = 0,
  dt2 = 0,
  by = geo_value,
  use = "na.or.complete",
  method = c("pearson", "kendall", "spearman")
)

Arguments

x
The epi_tibble object.

var1, var2
The variables in x to correlate.

dt1, dt2
Time shifts to consider for the two variables, respectively, before computing correlations. Negative shifts translate into in a lag value and positive shifts into a lead value; for example, if dt = -1, then the new value on June 2 is the original value on June 1; if dt = 1, then the new value on June 2 is the original value on June 3; if dt = 0, then the values are left as is. Default is 0 for both dt1 and dt2.

by
The variable(s) to group by, for the correlation computation. If geo_value, the default, then correlations are computed for each geo value, over all time; if time_value, then correlations are computed for each time, over all geo values. A grouping can also be any specified using number of columns of x; for example, we can use by = c(geo_value, age_group), assuming x has a column age_group, in order to compute correlations for each pair of geo value and age group. To omit a grouping entirely, use by = NULL.

use, method
Arguments to pass to cor(), with "na.or.complete" the default for use (different than cor()) and "pearson" the default for method (same as cor()).

Value

An tibble with the grouping columns first (geo_value, time_value, or possibly others), and then a column cor, which gives the correlation.