Creates an epi_df object

Creates a new epi_df object. By default, builds an empty tibble with the correct metadata for an epi_df object (ie. geo_type, time_type, and as_of). Refer to the below info. about the arguments for more details.

Usage

new_epi_df(
  x = tibble::tibble(),
  geo_type,
  time_type,
  as_of,
  additional_metadata = list()
)

Arguments

x

A data.frame, tibble::tibble, or tsibble::tsibble to be converted

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

as_of

Time value representing the time at which the given data were available. For example, if as_of is January 31, 2022, then the epi_df object that is created would represent the most up-to-date version of the data available as of January 31, 2022. If the as_of argument is missing, then the current day-time will be used.

additional_metadata

List of additional metadata to attach to the epi_df object. The metadata will have geo_type, time_type, and as_of fields; named entries from the passed list will be included as well. If your tibble has additional keys, be sure to specify them as a character vector in the other_keys component of additional_metadata.

Value

An epi_df object.