Subset of daily doctor visits and cases in archive format

This data source is based on information about outpatient visits, provided to us by health system partners, and also contains confirmed COVID-19 cases based on reports made available by the Center for Systems Science and Engineering at Johns Hopkins University. This example data ranges from June 1, 2020 to Dec 1, 2021, and is also limited to California, Florida, Texas, and New York.

Usage

archive_cases_dv_subset

Format

An epi_archive data format. The data table DT has 129,638 rows and 5 columns:

geo_value

the geographic value associated with each row of measurements.

time_value

the time value associated with each row of measurements.

version

the time value specifying the version for each row of measurements.

percent_cli

percentage of doctor’s visits with CLI (COVID-like illness) computed from medical insurance claims

case_rate_7d_av

7-day average signal of number of new confirmed deaths due to COVID-19 per 100,000 population, daily

Source

This object contains a modified part of the COVID-19 Data Repository by the Center for Systems Science and Engineering (CSSE) at Johns Hopkins University as republished in the COVIDcast Epidata API. This data set is licensed under the terms of the Creative Commons Attribution 4.0 International license by Johns Hopkins University on behalf of its Center for Systems Science in Engineering. Copyright Johns Hopkins University 2020.

Modifications:

From the COVIDcast Doctor Visits API: The signal percent_cli is taken directly from the API without changes.

From the COVIDcast Epidata API: case_rate_7d_av signal was computed by Delphi from the original JHU-CSSE data by calculating moving averages of the preceding 7 days, so the signal for June 7 is the average of the underlying data for June 1 through 7, inclusive.

Furthermore, the data is a subset of the full dataset, the signal names slightly altered, and formatted into a tibble.