Convert to epi_archive format

Converts a data frame, data table, or tibble into an epi_archive object. See the archive vignette for examples.

as_epi_archive(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata = list()
)

Arguments

x
A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type
Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type
Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys
Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata
List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

Value

An epi_archive object.

Details

This simply a wrapper around the new() method of the epi_archive class, so for example:

x &lt;- as_epi_archive(df, geo_type = "state", time_type = "day")

would be equivalent to:

x &lt;- epi_archive$new(df, geo_type = "state", time_type = "day")