Convert to epi_archive format

Converts a data frame, data table, or tibble into an epi_archive object. See the archive vignette for examples.

Usage

as_epi_archive(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata = list()
)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

Value

An epi_archive object.

Details

This simply a wrapper around the new() method of the epi_archive class, so for example:

x <- as_epi_archive(df, geo_type = "state", time_type = "day")

would be equivalent to:

x <- epi_archive$new(df, geo_type = "state", time_type = "day")

Examples

# Simple ex. with necessary keys
tib <- tibble::tibble(
  geo_value = rep(c("ca", "hi"), each = 5),
  time_value = rep(seq(as.Date("2020-01-01"), 
                       by = 1, length.out = 5), times = 2),
  version = rep(seq(as.Date("2020-01-02"), 
                    by = 1, length.out = 5), times = 2),
  value = rnorm(10, mean = 2, sd = 1)
)

toy_epi_archive <- tib %>% as_epi_archive(geo_type = "state", 
                                          time_type = "day")
toy_epi_archive 
#> An `epi_archive` object, with metadata:
#> * geo_type  = state
#> * time_type = day
#> ----------
#> * min time value = 2020-01-01
#> * max time value = 2020-01-05
#> * min version    = 2020-01-02
#> * max version    = 2020-01-06
#> ----------
#> Data archive (stored in DT field): 10 x 4
#> ----------
#> Columns in DT: geo_value, time_value, version, value
#> ----------
#> Public methods: initialize, print, as_of, merge, slide, clone

# Ex. with an additional key for county
df <- data.frame (geo_value  = c(replicate(2, "ca"), replicate(2, "fl")),
                 county = c(1, 3, 2, 5),
                 time_value = c("2020-06-01",
                 "2020-06-02",
                 "2020-06-01",
                 "2020-06-02"),
                 version = c("2020-06-02",
                 "2020-06-03",
                 "2020-06-02",
                 "2020-06-03"),
                 cases = c(1, 2, 3, 4),
                 cases_rate = c(0.01, 0.02, 0.01, 0.05))

x <- df %>% as_epi_archive(geo_type = "state",
                          time_type = "day",
                          other_keys = "county")