Convert to epi_archive format

Converts a data frame, data table, or tibble into an epi_archive object. See the archive vignette for examples. The parameter descriptions below are copied from there

Usage

as_epi_archive(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata = list(),
  compactify = NULL,
  clobberable_versions_start = NA,
  versions_end = max_version_with_row_in(x)
)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

compactify

Optional; Boolean or NULL: should we remove rows that are considered redundant for the purposes of epi_archive's built-in methods such as as_of? As these methods use the last version of each observation carried forward (LOCF) to interpolate between the version data provided, rows that don't change these LOCF results can potentially be omitted to save space. TRUE will remove these rows, FALSE will not, and missing or NULL will remove these rows and issue a warning. Generally, this can be set to TRUE, but if you directly inspect or edit the fields of the epi_archive such as its DT, you will have to determine whether compactify=TRUE will produce the desired results. If compactification here is removing a large proportion of the rows, this may indicate a potential for space, time, or bandwidth savings upstream the data pipeline, e.g., when fetching, storing, or preparing the input data x

clobberable_versions_start

Optional; length-1; either a value of the same class and typeof as x$version, or an NA of any class and typeof: specifically, either (a) the earliest version that could be subject to "clobbering" (being overwritten with different update data, but using the same version tag as the old update data), or (b) NA, to indicate that no versions are clobberable. There are a variety of reasons why versions could be clobberable under routine circumstances, such as (a) today's version of one/all of the columns being published after initially being filled with NA or LOCF, (b) a buggy version of today's data being published but then fixed and republished later in the day, or (c) data pipeline delays (e.g., publisher uploading, periodic scraping, database syncing, periodic fetching, etc.) that make events (a) or (b) reflected later in the day (or even on a different day) than expected; potential causes vary between different data pipelines. The default value is NA, which doesn't consider any versions to be clobberable. Another setting that may be appropriate for some pipelines is max_version_with_row_in(x).

versions_end

Optional; length-1, same class and typeof as x$version: what is the last version we have observed? The default is max_version_with_row_in(x), but values greater than this could also be valid, and would indicate that we observed additional versions of the data beyond max(x$version), but they all contained empty updates. (The default value of clobberable_versions_start does not fully trust these empty updates, and assumes that any version >= max(x$version) could be clobbered.) If nrow(x) == 0, then this argument is mandatory.

Value

An epi_archive object.

Details

This simply a wrapper around the new() method of the epi_archive class, so for example:

x <- as_epi_archive(df, geo_type = "state", time_type = "day")

would be equivalent to:

x <- epi_archive$new(df, geo_type = "state", time_type = "day")

Examples

# Simple ex. with necessary keys
tib <- tibble::tibble(
  geo_value = rep(c("ca", "hi"), each = 5),
  time_value = rep(seq(as.Date("2020-01-01"),
    by = 1, length.out = 5
  ), times = 2),
  version = rep(seq(as.Date("2020-01-02"),
    by = 1, length.out = 5
  ), times = 2),
  value = rnorm(10, mean = 2, sd = 1)
)

toy_epi_archive <- tib %>% as_epi_archive(
  geo_type = "state",
  time_type = "day"
)
toy_epi_archive
#> → An `epi_archive` object, with metadata:
#> ℹ Min/max time values: 2020-01-01 / 2020-01-05
#> ℹ First/last version with update: 2020-01-02 / 2020-01-06
#> ℹ Versions end: 2020-01-06
#> ℹ Public R6 methods: initialize, print, as_of, fill_through_version,
#>   truncate_versions_after, merge, group_by, slide, and clone
#> ℹ A preview of the table (10 rows x 4 columns):
#> Key: <geo_value, time_value, version>
#>     geo_value time_value    version      value
#>        <char>     <Date>     <Date>      <num>
#>  1:        ca 2020-01-01 2020-01-02  0.5999565
#>  2:        ca 2020-01-02 2020-01-03  2.2553171
#>  3:        ca 2020-01-03 2020-01-04 -0.4372636
#>  4:        ca 2020-01-04 2020-01-05  1.9944287
#>  5:        ca 2020-01-05 2020-01-06  2.6215527
#>  6:        hi 2020-01-01 2020-01-02  3.1484116
#>  7:        hi 2020-01-02 2020-01-03  0.1781823
#>  8:        hi 2020-01-03 2020-01-04  1.7526747
#>  9:        hi 2020-01-04 2020-01-05  1.7558004
#> 10:        hi 2020-01-05 2020-01-06  1.7172946

# Ex. with an additional key for county
df <- data.frame(
  geo_value = c(replicate(2, "ca"), replicate(2, "fl")),
  county = c(1, 3, 2, 5),
  time_value = c(
    "2020-06-01",
    "2020-06-02",
    "2020-06-01",
    "2020-06-02"
  ),
  version = c(
    "2020-06-02",
    "2020-06-03",
    "2020-06-02",
    "2020-06-03"
  ),
  cases = c(1, 2, 3, 4),
  cases_rate = c(0.01, 0.02, 0.01, 0.05)
)

x <- df %>% as_epi_archive(
  geo_type = "state",
  time_type = "day",
  other_keys = "county"
)