Convert to epi_df format

Converts a data frame or tibble into an epi_df object. See the getting started guide for examples.

Usage

as_epi_df(x, ...)

# S3 method for epi_df
as_epi_df(x, ...)

# S3 method for tbl_df
as_epi_df(x, geo_type, time_type, as_of, additional_metadata = list(), ...)

# S3 method for data.frame
as_epi_df(x, geo_type, time_type, as_of, additional_metadata = list(), ...)

# S3 method for tbl_ts
as_epi_df(x, geo_type, time_type, as_of, additional_metadata = list(), ...)

Arguments

x

A data.frame, tibble::tibble, or tsibble::tsibble to be converted

...

Additional arguments passed to methods.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

as_of

Time value representing the time at which the given data were available. For example, if as_of is January 31, 2022, then the epi_df object that is created would represent the most up-to-date version of the data available as of January 31, 2022. If the as_of argument is missing, then the current day-time will be used.

additional_metadata

List of additional metadata to attach to the epi_df object. The metadata will have geo_type, time_type, and as_of fields; named entries from the passed list or will be included as well.

Value

An epi_df object.

Methods (by class)

as_epi_df(epi_df): Simply returns the epi_df object unchanged.

as_epi_df(tbl_df): The input tibble x must contain the columns geo_value and time_value. All other columns will be preserved as is, and treated as measured variables. If as_of is missing, then the function will try to guess it from an as_of, issue, or version column of x (if any of these are present), or from as an as_of field in its metadata (stored in its attributes); if this fails, then the current day-time will be used.

as_epi_df(data.frame): Works analogously to as_epi_df.tbl_df().

as_epi_df(tbl_ts): Works analogously to as_epi_df.tbl_df(), except that the tbl_ts class is dropped, and any key variables (other than "geo_value") are added to the metadata of the returned object, under the other_keys field.

Examples

# Convert a `tsibble` that has county code as an extra key
# Notice that county code should be a character string to preserve any leading zeroes

ex1_input <- tibble::tibble(
  geo_value = rep(c("ca", "fl", "pa"), each = 3),
  county_code = c("06059","06061","06067",
                  "12111","12113","12117",
                  "42101", "42103","42105"),
  time_value = rep(seq(as.Date("2020-06-01"), as.Date("2020-06-03"),
                       by = "day"), length.out = length(geo_value)),
  value = 1:length(geo_value) + 0.01 * rnorm(length(geo_value))
) %>% 
  tsibble::as_tsibble(index = time_value, key = c(geo_value, county_code))

# The `other_keys` metadata (`"county_code"` in this case) is automatically
# inferred from the `tsibble`'s `key`:
ex1 <- as_epi_df(x = ex1_input, geo_type = "state", time_type = "day", as_of = "2020-06-03")
attr(ex1,"metadata")[["other_keys"]]
#> [1] "county_code"

# Dealing with misspecified column names:
# Geographical and temporal information must be provided in columns named
# `geo_value` and `time_value`; if we start from a data frame with a
# different format, it must be converted to use `geo_value` and `time_value`
# before calling `as_epi_df`.

ex2_input <- tibble::tibble(
  state = rep(c("ca", "fl", "pa"), each = 3), # misnamed
  pol = rep(c("blue", "swing", "swing"), each = 3), # extra key
  reported_date = rep(seq(as.Date("2020-06-01"), as.Date("2020-06-03"),
                          by = "day"), length.out = length(state)), # misnamed
  value = 1:length(state) + 0.01 * rnorm(length(state))
) 

print(ex2_input)
#> # A tibble: 9 × 4
#>   state pol   reported_date value
#>   <chr> <chr> <date>        <dbl>
#> 1 ca    blue  2020-06-01     1.01
#> 2 ca    blue  2020-06-02     2.01
#> 3 ca    blue  2020-06-03     2.98
#> 4 fl    swing 2020-06-01     4.01
#> 5 fl    swing 2020-06-02     5.00
#> 6 fl    swing 2020-06-03     6.02
#> 7 pa    swing 2020-06-01     6.99
#> 8 pa    swing 2020-06-02     8.01
#> 9 pa    swing 2020-06-03     9.01

ex2 <- ex2_input %>% dplyr::rename(geo_value = state, time_value = reported_date) %>%
  as_epi_df(geo_type = "state", as_of = "2020-06-03", 
            additional_metadata = c(other_keys = "pol"))

attr(ex2,"metadata")
#> $geo_type
#> [1] "state"
#> 
#> $time_type
#> [1] "day"
#> 
#> $as_of
#> [1] "2020-06-03"
#> 
#> $other_keys
#> [1] "pol"
#> 

# Adding additional keys to an `epi_df` object

ex3_input <- jhu_csse_county_level_subset %>%
  dplyr::filter(time_value > "2021-12-01", state_name == "Massachusetts") %>%
  dplyr::slice_tail(n = 6) 

ex3 <- ex3_input %>% 
  tsibble::as_tsibble() %>% # needed to add the additional metadata
  dplyr::mutate(state = rep("MA",6)) %>%
  as_epi_df(additional_metadata = c(other_keys = "state"))

attr(ex3,"metadata")
#> $geo_type
#> [1] "county"
#> 
#> $time_type
#> [1] "day"
#> 
#> $as_of
#> [1] "2022-08-01 01:53:46 UTC"
#> 
#> $other_keys
#> [1] "state"
#>