epi_archive object

An epi_archive is an R6 class which contains a data table along with several relevant pieces of metadata. The data table can be seen as the full archive (version history) for some signal variables of interest.

Details

An epi_archive is an R6 class which contains a data table DT, of class data.table from the data.table package, with (at least) the following columns:

geo_value: the geographic value associated with each row of measurements.

time_value: the time value associated with each row of measurements.

version: the time value specifying the version for each row of measurements. For example, if in a given row the version is January 15, 2022 and time_value is January 14, 2022, then this row contains the measurements of the data for January 14, 2022 that were available one day later.

The data table DT has key variables geo_value, time_value, version, as well as any other specified in the metadata (described below). There can only be a single row per unique combination of key variables, and thus the key variables are critical for figuring out how to generate a snapshot of data from the archive, as of a given version.

In general, last-observation-carried-forward (LOCF) is used to data in between recorded versions. Currently, deletions must be represented as revising a row to a special state (e.g., making the entries NA or including a special column that flags the data as removed and performing some kind of post-processing), and the archive is unaware of what this state is.

A word of caution: R6 objects, unlike most other objects in R, have reference semantics. A primary consequence of this is that objects are not copied when modified. You can read more about this in Hadley Wickham's Advanced R book.

Metadata

The following pieces of metadata are included as fields in an epi_archive object:

geo_type: the type for the geo values.

time_type: the type for the time values.

max_version: the max version in the data archive.

other_keys: the names of the other key variables (apart from "geo_value", "time_value", and "version") in the data archive.

additional_metadata: list of additional metadata for the data archive.

Unlike an epi_df object, metadata for an epi_archive object x can be accessed (and altered) directly via x$geo_type.

Generating Snapshots

An epi_archive object can be used to generate a snapshot of the data in epi_df format, which represents the most up-to-date values of the signal variables, as of the specified version. This is accomplished by calling the as_of() method for an epi_archive object x, for example:

x$as_of(as.Date("2022-01-15"))

to generate a snapshot as of January 15, 2022. More details on the as_of() method are documented below.

See also

Methods

Public methods

epi_archive$new()

epi_archive$as_of()

epi_archive$clone()

Method new()

Creates a new epi_archive object.

Usage

epi_archive$new(
  x,
  geo_type,
  time_type,
  max_version,
  other_keys,
  additional_metadata
)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

max_version

Maximum version in the data archive. If missing, then the function will set it equal to the maximum of the version column in x.

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

Returns

An epi_archive object.

Method as_of()

Generates a snapshot in epi_df format as of a given version.

Usage

epi_archive$as_of(max_version, min_time_value = -Inf)

Arguments

max_version

Time value specifying the max version to permit in the snapshot. That is, the snapshot will comprise the unique rows that represent the most up-to-date signal values, as of the specified max_version, whose time values are at least min_time_value.

min_time_value

Time value specifying the min time value to permit in the snapshot. Default is -Inf, which effectively means that there is no minimum considered.

Returns

An epi_df object.

Method clone()

The objects of this class are cloneable with this method.

Usage

epi_archive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.