epi_archive object

An epi_archive is an R6 class which contains a data table along with several relevant pieces of metadata. The data table can be seen as the full archive (version history) for some signal variables of interest.

Details

An epi_archive is an R6 class which contains a data table DT, of class data.table from the data.table package, with (at least) the following columns:

geo_value: the geographic value associated with each row of measurements.

time_value: the time value associated with each row of measurements.

version: the time value specifying the version for each row of measurements. For example, if in a given row the version is January 15, 2022 and time_value is January 14, 2022, then this row contains the measurements of the data for January 14, 2022 that were available one day later.

The data table DT has key variables geo_value, time_value, version, as well as any others (these can be specified when instantiating the epi_archive object via the other_keys argument, and/or set by operating on DT directly). There can only be a single row per unique combination of key variables, and thus the key variables are critical for figuring out how to generate a snapshot of data from the archive, as of a given version.

In general, last observation carried forward (LOCF) is used to data in between recorded versions. Currently, deletions must be represented as revising a row to a special state (e.g., making the entries NA or including a special column that flags the data as removed and performing some kind of post-processing), and the archive is unaware of what this state is.

A word of caution: R6 objects, unlike most other objects in R, have reference semantics. A primary consequence of this is that objects are not copied when modified. You can read more about this in Hadley Wickham's Advanced R book.

Metadata

The following pieces of metadata are included as fields in an epi_archive object:

geo_type: the type for the geo values.

time_type: the type for the time values.

additional_metadata: list of additional metadata for the data archive.

Unlike an epi_df object, metadata for an epi_archive object x can be accessed (and altered) directly, as in x$geo_type or x$time_type, etc. Like an epi_df object, the geo_type and time_type fields in the metadata of an epi_archive object are not currently used by any downstream functions in the epiprocess package, and serve only as useful bits of information to convey about the data set at hand.

Generating Snapshots

An epi_archive object can be used to generate a snapshot of the data in epi_df format, which represents the most up-to-date values of the signal variables, as of the specified version. This is accomplished by calling the as_of() method for an epi_archive object x. More details on this method are documented in the wrapper function epix_as_of().

Sliding Computations

We can run a sliding computation over an epi_archive object, much like epi_slide() does for an epi_df object. This is accomplished by calling the slide() method for an epi_archive object, which works similarly to the way epi_slide() works for an epi_df object, but with one key difference: it is version-aware. That is, for an epi_archive object, the sliding computation at any given reference time point t is performed on data that would have been available as of t. More details on slide() are documented in the wrapper function epix_slide().

Methods

Public methods

epi_archive$new()

epi_archive$print()

epi_archive$as_of()

epi_archive$merge()

epi_archive$slide()

epi_archive$clone()

Method new()

Creates a new epi_archive object.

Usage

epi_archive$new(x, geo_type, time_type, other_keys, additional_metadata)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

Returns

An epi_archive object.

Method print()

Usage

epi_archive$print()

Method as_of()

Generates a snapshot in epi_df format as of a given version. See the documentation for the wrapper function epix_as_of() for details.

Usage

epi_archive$as_of(max_version, min_time_value = -Inf)

Method merge()

Merges another data.table with the current one, and allows for a post-filling of NA values by last observation carried forward (LOCF). See the documentation for the wrapper function epix_merge() for details.

Usage

epi_archive$merge(y, ..., locf = TRUE, nan = NA)

Method slide()

Slides a given function over variables in an epi_archive object. See the documentation for the wrapper function epix_as_of() for details.

Usage

epi_archive$slide(
  f,
  ...,
  n = 7,
  group_by,
  ref_time_values,
  time_step,
  complete = FALSE,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_rows = FALSE
)

Method clone()

The objects of this class are cloneable with this method.

Usage

epi_archive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.