Archive (data version history) for an epi_df object

Archive (data version history) for an epi_df object

Archive (data version history) for an epi_df object

Details

Contains version history for an epi_df object, and enables fast querying of snapshots of the epi_df object as of certain "issues" (versions). See the data versioning vignette for examples.

Version history can be input as a data frame combining full snapshots of the epi_df as of several issue times, or using only the newly added or revised rows for each issue, or using some combination of these two (including "updates" for things that didn't actually change). Last-observation-carried-forward (LOCF) is used to data in between recorded updates. Currently, deletions must be represented as revising a row to a special state (e.g., making the entries NA or including a special column that flags the data as removed and performing post-processing), and the archive is unaware of what this state is.

Methods

Public methods

epi_archive$new()

epi_archive$epi_df_as_of()

epi_archive$naming_info()

epi_archive$max_issue()

epi_archive$issues_with_updates()

epi_archive$update_DT_as_of()

epi_archive$clone()

Method new()

Create a new epi_archive with the given update data

Usage

epi_archive$new(
  update.df,
  issue.colname = "issue",
  geo.colname = "geo_value",
  time.colname = "time_value",
  other.key.colnames = character(0L),
  max.issue = max(update.df[[issue.colname]])
)

Arguments

update.df

the update data

issue.colname

name of the column with the issue time of the corresponding updates; operations such as sort, <=, and max must make sense on this column, with earlier issues "less than" later issues

geo.colname

the name of the column that will become geo_value in the epi_df

time.colname

the name of the column that will become time_value in the epi_df

other.key.colnames

the names of any other columns that would be used to index a single measurement in this data set, such as the age group the measurement corresponds to (if the data set includes an age group breakdown); there should only be a single row per issue-geo-time-other-key-cols combination.

max.issue

the latest issue for which update data was available; defaults to the maximum issue time in the update.df, but if there were no additions or revisions in subsequent issues, it could be later. However, due to details regarding database replica syncing times in upstream APIs, using the default might be safer than whatever we think the max issue should be.

Returns

an epi_archive object

Method epi_df_as_of()

Get the epi_df as of some issue time

Usage

epi_archive$epi_df_as_of(issue)

Arguments

issue

the desired as-of issue time

Returns

an epi_df with data as of the specified issue time, issue recorded in the metadata, the geo column renamed to geo_value and time column to time_value, and the other key colnames recorded in the metadata

Method naming_info()

Return the name settings in a list

Usage

epi_archive$naming_info()

Method max_issue()

Return the max issue value recorded by this archive (whether it had updates or not)

Usage

epi_archive$max_issue()

Method issues_with_updates()

Return the issue values for which updates are recorded in this archive (that is, whether they had updates in the data frame used to form this archive, regardless of whether those "updates" actually added or revised any data)

Usage

epi_archive$issues_with_updates()

Method update_DT_as_of()

Return the recorded update data up through the given issue value, inside a data.table object which is fine to modify without copying.

Usage

epi_archive$update_DT_as_of(issue)

Arguments

issue

the max issue value that should appear in the result

Method clone()

The objects of this class are cloneable with this method.

Usage

epi_archive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.