epi_archive object

An epi_archive is an R6 class which contains a data table along with several relevant pieces of metadata. The data table can be seen as the full archive (version history) for some signal variables of interest.

Details

An epi_archive is an R6 class which contains a data table DT, of class data.table from the data.table package, with (at least) the following columns:

geo_value: the geographic value associated with each row of measurements.

time_value: the time value associated with each row of measurements.

version: the time value specifying the version for each row of measurements. For example, if in a given row the version is January 15, 2022 and time_value is January 14, 2022, then this row contains the measurements of the data for January 14, 2022 that were available one day later.

The data table DT has key variables geo_value, time_value, version, as well as any others (these can be specified when instantiating the epi_archive object via the other_keys argument, and/or set by operating on DT directly). Refer to the documentation for as_epi_archive() for information and examples of relevant parameter names for an epi_archive object. Note that there can only be a single row per unique combination of key variables, and thus the key variables are critical for figuring out how to generate a snapshot of data from the archive, as of a given version.

In general, the last version of each observation is carried forward (LOCF) to fill in data between recorded versions, and between the last recorded update and the versions_end. One consequence is that the DT doesn't have to contain a full snapshot of every version (although this generally works), but can instead contain only the rows that are new or changed from the previous version (see compactify, which does this automatically). Currently, deletions must be represented as revising a row to a special state (e.g., making the entries NA or including a special column that flags the data as removed and performing some kind of post-processing), and the archive is unaware of what this state is. Note that NAs can be introduced by epi_archive methods for other reasons, e.g., in epix_fill_through_version and epix_merge, if requested, to represent potential update data that we do not yet have access to; or in epix_merge to represent the "value" of an observation before the version in which it was first released, or if no version of that observation appears in the archive data at all.

A word of caution: R6 objects, unlike most other objects in R, have reference semantics. A primary consequence of this is that objects are not copied when modified. You can read more about this in Hadley Wickham's Advanced R book. In order to construct a modified archive while keeping the original intact, first make a clone using the $clone method, then overwrite the clone's DT field with data.table::copy(clone$DT), and finally perform the modifications on the clone.

Metadata

The following pieces of metadata are included as fields in an epi_archive object:

geo_type: the type for the geo values.

time_type: the type for the time values.

additional_metadata: list of additional metadata for the data archive.

Unlike an epi_df object, metadata for an epi_archive object x can be accessed (and altered) directly, as in x$geo_type or x$time_type, etc. Like an epi_df object, the geo_type and time_type fields in the metadata of an epi_archive object are not currently used by any downstream functions in the epiprocess package, and serve only as useful bits of information to convey about the data set at hand.

Generating Snapshots

An epi_archive object can be used to generate a snapshot of the data in epi_df format, which represents the most up-to-date values of the signal variables, as of the specified version. This is accomplished by calling the as_of() method for an epi_archive object x. More details on this method are documented in the wrapper function epix_as_of().

Sliding Computations

We can run a sliding computation over an epi_archive object, much like epi_slide() does for an epi_df object. This is accomplished by calling the slide() method for an epi_archive object, which works similarly to the way epi_slide() works for an epi_df object, but with one key difference: it is version-aware. That is, for an epi_archive object, the sliding computation at any given reference time point t is performed on data that would have been available as of t. More details on slide() are documented in the wrapper function epix_slide().

Methods

Public methods

epi_archive$new()

epi_archive$print()

epi_archive$as_of()

epi_archive$fill_through_version()

epi_archive$truncate_versions_after()

epi_archive$merge()

epi_archive$group_by()

epi_archive$slide()

epi_archive$clone()

Method new()

Creates a new epi_archive object.

Usage

epi_archive$new(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata,
  compactify,
  clobberable_versions_start,
  versions_end
)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

compactify

Optional; Boolean or NULL: should we remove rows that are considered redundant for the purposes of epi_archive's built-in methods such as as_of? As these methods use the last version of each observation carried forward (LOCF) to interpolate between the version data provided, rows that don't change these LOCF results can potentially be omitted to save space while maintaining the same behavior (with the help of the clobberable_versions_start and versions_end fields in some edge cases). TRUE will remove these rows, FALSE will not, and missing or NULL will remove these rows and issue a warning. Generally, this can be set to TRUE, but if you directly inspect or edit the fields of the epi_archive such as its DT, or rely on redundant updates to achieve a certain behavior of the ref_time_values default in epix_slide, you will have to determine whether compactify=TRUE will produce the desired results. If compactification here is removing a large proportion of the rows, this may indicate a potential for space, time, or bandwidth savings upstream the data pipeline, e.g., by avoiding fetching, storing, or processing these rows of x.

clobberable_versions_start

Optional; as in as_epi_archive

versions_end

Optional; as in as_epi_archive

Details

Refer to the documentation for as_epi_archive() for more information and examples of parameter names.

Returns

An epi_archive object.

Method print()

Usage

epi_archive$print(class = TRUE, methods = TRUE)

Method as_of()

Generates a snapshot in epi_df format as of a given version. See the documentation for the wrapper function epix_as_of() for details.

Usage

epi_archive$as_of(max_version, min_time_value = -Inf, all_versions = FALSE)

Method fill_through_version()

Fill in unobserved history using requested scheme by mutating self and potentially reseating its fields. See epix_fill_through_version for a full description of the non-R6-method version, which doesn't mutate the input archive but might alias its fields.

Usage

epi_archive$fill_through_version(fill_versions_end, how = c("na", "locf"))

Arguments

fill_versions_end

as in epix_fill_through_version

how

as in epix_fill_through_version

Method truncate_versions_after()

Filter to keep only older versions, mutating the archive by potentially reseating but not mutating some fields. DT is likely, but not guaranteed, to be copied. Returns the mutated archive invisibly.

Usage

epi_archive$truncate_versions_after(max_version)

Arguments

max_version

as in epix_truncate_versions_after

x

as in epix_truncate_versions_after

Method merge()

Merges another epi_archive with the current one, mutating the current one by reseating its DT and several other fields, but avoiding mutation of the old DT; returns the current archive invisibly. See epix_merge for a full description of the non-R6-method version, which does not mutate either archive, and does not alias either archive's DT.

Usage

epi_archive$merge(
  y,
  sync = c("forbid", "na", "locf", "truncate"),
  compactify = TRUE
)

Arguments

y

as in epix_merge

sync

as in epix_merge

compactify

as in epix_merge

Method group_by()

Usage

epi_archive$group_by(
  ...,
  .add = FALSE,
  .drop = dplyr::group_by_drop_default(self)
)

Method slide()

Slides a given function over variables in an epi_archive object. See the documentation for the wrapper function epix_slide() for details.

Usage

epi_archive$slide(
  f,
  ...,
  before,
  ref_time_values,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_versions = FALSE
)

Method clone()

The objects of this class are cloneable with this method.

Usage

epi_archive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

tib <- tibble::tibble(
  geo_value = rep(c("ca", "hi"), each = 5),
  time_value = rep(seq(as.Date("2020-01-01"), 
                       by = 1, length.out = 5), times = 2),
  version = rep(seq(as.Date("2020-01-02"), 
                    by = 1, length.out = 5), times = 2),
  value = rnorm(10, mean = 2, sd = 1)
)

toy_epi_archive <- tib %>% epi_archive$new(geo_type = "state", 
                                           time_type = "day")
toy_epi_archive 
#> An `epi_archive` object, with metadata:
#> * geo_type  = state
#> * time_type = day
#> ----------
#> * min time value = 2020-01-01
#> * max time value = 2020-01-05
#> * first version with update = 2020-01-02
#> * last version with update = 2020-01-06
#> * No clobberable versions
#> * versions end   = 2020-01-06
#> ----------
#> Data archive (stored in DT field): 10 x 4
#> Columns in DT: geo_value, time_value, version, value
#> ----------
#> Public R6 methods: initialize, print, as_of, fill_through_version, 
#>                    truncate_versions_after, merge, group_by, slide, clone