epi_archive object

An epi_archive is an R6 class which contains a data table along with several relevant pieces of metadata. The data table can be seen as the full archive (version history) for some signal variables of interest.

Details

An epi_archive is an R6 class which contains a data table DT, of class data.table from the data.table package, with (at least) the following columns:

geo_value: the geographic value associated with each row of measurements.

time_value: the time value associated with each row of measurements.

version: the time value specifying the version for each row of measurements. For example, if in a given row the version is January 15, 2022 and time_value is January 14, 2022, then this row contains the measurements of the data for January 14, 2022 that were available one day later.

The data table DT has key variables geo_value, time_value, version, as well as any others (these can be specified when instantiating the epi_archive object via the other_keys argument, and/or set by operating on DT directly). Refer to the documentation for as_epi_archive() for information and examples of relevant parameter names for an epi_archive object. Note that there can only be a single row per unique combination of key variables, and thus the key variables are critical for figuring out how to generate a snapshot of data from the archive, as of a given version.

In general, the last version of each observation is carried forward (LOCF) to fill in data between recorded versions, and between the last recorded update and the versions_end. One consequence is that the DT doesn't have to contain a full snapshot of every version (although this generally works), but can instead contain only the rows that are new or changed from the previous version (see compactify, which does this automatically). Currently, deletions must be represented as revising a row to a special state (e.g., making the entries NA or including a special column that flags the data as removed and performing some kind of post-processing), and the archive is unaware of what this state is. Note that NAs can be introduced by epi_archive methods for other reasons, e.g., in epix_fill_through_version and epix_merge, if requested, to represent potential update data that we do not yet have access to; or in epix_merge to represent the "value" of an observation before the version in which it was first released, or if no version of that observation appears in the archive data at all.

A word of caution: R6 objects, unlike most other objects in R, have reference semantics. A primary consequence of this is that objects are not copied when modified. You can read more about this in Hadley Wickham's Advanced R book. In order to construct a modified archive while keeping the original intact, first make a clone using the $clone method, then overwrite the clone's DT field with data.table::copy(clone$DT), and finally perform the modifications on the clone.

Metadata

The following pieces of metadata are included as fields in an epi_archive object:

geo_type: the type for the geo values.

time_type: the type for the time values.

additional_metadata: list of additional metadata for the data archive.

Unlike an epi_df object, metadata for an epi_archive object x can be accessed (and altered) directly, as in x$geo_type or x$time_type, etc. Like an epi_df object, the geo_type and time_type fields in the metadata of an epi_archive object are not currently used by any downstream functions in the epiprocess package, and serve only as useful bits of information to convey about the data set at hand.

Generating Snapshots

An epi_archive object can be used to generate a snapshot of the data in epi_df format, which represents the most up-to-date values of the signal variables, as of the specified version. This is accomplished by calling the as_of() method for an epi_archive object x. More details on this method are documented in the wrapper function epix_as_of().

Sliding Computations

We can run a sliding computation over an epi_archive object, much like epi_slide() does for an epi_df object. This is accomplished by calling the slide() method for an epi_archive object, which works similarly to the way epi_slide() works for an epi_df object, but with one key difference: it is version-aware. That is, for an epi_archive object, the sliding computation at any given reference time point t is performed on data that would have been available as of t. More details on slide() are documented in the wrapper function epix_slide().

Public fields

DT

(data.table)
the (optionally compactified) datatable

geo_type

(string)
the resolution of the geographic label (e.g. state)

time_type

(string)
the resolution of the time column (e.g. day)

additional_metadata

(named list)
any extra fields, such as other_keys

clobberable_versions_start

(length-1 of same type&class as version column, or NA)
the earliest version number that might be rewritten in the future without assigning a new version date/number, or NA if this won't happen

versions_end

(length-1 of same type&class as version column)
the latest version observed

Methods

Public methods

epi_archive$new()

epi_archive$print()

epi_archive$as_of()

epi_archive$fill_through_version()

epi_archive$truncate_versions_after()

epi_archive$merge()

epi_archive$group_by()

epi_archive$slide()

epi_archive$clone()

Method new()

Creates a new epi_archive object.

Usage

epi_archive$new(
  x,
  geo_type,
  time_type,
  other_keys,
  additional_metadata,
  compactify,
  clobberable_versions_start,
  versions_end
)

Arguments

x

A data frame, data table, or tibble, with columns geo_value, time_value, version, and then any additional number of columns.

geo_type

Type for the geo values. If missing, then the function will attempt to infer it from the geo values present; if this fails, then it will be set to "custom".

time_type

Type for the time values. If missing, then the function will attempt to infer it from the time values present; if this fails, then it will be set to "custom".

other_keys

Character vector specifying the names of variables in x that should be considered key variables (in the language of data.table) apart from "geo_value", "time_value", and "version".

additional_metadata

List of additional metadata to attach to the epi_archive object. The metadata will have geo_type and time_type fields; named entries from the passed list or will be included as well.

compactify

Optional; Boolean or NULL: should we remove rows that are considered redundant for the purposes of epi_archive's built-in methods such as as_of? As these methods use the last version of each observation carried forward (LOCF) to interpolate between the version data provided, rows that don't change these LOCF results can potentially be omitted to save space while maintaining the same behavior (with the help of the clobberable_versions_start and versions_end fields in some edge cases). TRUE will remove these rows, FALSE will not, and missing or NULL will remove these rows and issue a warning. Generally, this can be set to TRUE, but if you directly inspect or edit the fields of the epi_archive such as its DT, or rely on redundant updates to achieve a certain behavior of the ref_time_values default in epix_slide, you will have to determine whether compactify=TRUE will produce the desired results. If compactification here is removing a large proportion of the rows, this may indicate a potential for space, time, or bandwidth savings upstream the data pipeline, e.g., by avoiding fetching, storing, or processing these rows of x.

clobberable_versions_start

Optional; as in as_epi_archive

versions_end

Optional; as in as_epi_archive

Details

Refer to the documentation for as_epi_archive() for more information and examples of parameter names. Print information about an archive

Returns

An epi_archive object.

Method print()

Usage

epi_archive$print(class = TRUE, methods = TRUE)

Arguments

class

Boolean; whether to print the class label header

methods

Boolean; whether to print all available methods of the archive

Method as_of()

Generates a snapshot in epi_df format as of a given version. See the documentation for the wrapper function epix_as_of() for details. The parameter descriptions below are copied from there

Usage

epi_archive$as_of(max_version, min_time_value = -Inf, all_versions = FALSE)

Arguments

max_version

Version specifying the max version to permit in the snapshot. That is, the snapshot will comprise the unique rows of the current archive data that represent the most up-to-date signal values, as of the specified max_version (and whose time_values are at least min_time_value).

min_time_value

Time value specifying the min time_value to permit in the snapshot. Default is -Inf, which effectively means that there is no minimum considered.

all_versions

Boolean; If all_versions = TRUE, then the output will be in epi_archive format, and contain rows in the specified time_value range having version <= max_version. The resulting object will cover a potentially narrower version and time_value range than x, depending on user-provided arguments. Otherwise, there will be one row in the output for the max_version of each time_value. Default is FALSE.

x

An epi_archive object

Method fill_through_version()

Fill in unobserved history using requested scheme by mutating self and potentially reseating its fields. See epix_fill_through_version for a full description of the non-R6-method version, which doesn't mutate the input archive but might alias its fields.

Usage

epi_archive$fill_through_version(fill_versions_end, how = c("na", "locf"))

Arguments

fill_versions_end

as in epix_fill_through_version

how

as in epix_fill_through_version

Method truncate_versions_after()

Filter to keep only older versions, mutating the archive by potentially reseating but not mutating some fields. DT is likely, but not guaranteed, to be copied. Returns the mutated archive invisibly.

Usage

epi_archive$truncate_versions_after(max_version)

Arguments

max_version

as in epix_truncate_versions_after

x

as in epix_truncate_versions_after

Method merge()

Merges another epi_archive with the current one, mutating the current one by reseating its DT and several other fields, but avoiding mutation of the old DT; returns the current archive invisibly. See epix_merge for a full description of the non-R6-method version, which does not mutate either archive, and does not alias either archive's DT.

Usage

epi_archive$merge(
  y,
  sync = c("forbid", "na", "locf", "truncate"),
  compactify = TRUE
)

Arguments

y

as in epix_merge

sync

as in epix_merge

compactify

as in epix_merge group an epi_archive

Method group_by()

group an epi_archive

Usage

epi_archive$group_by(
  ...,
  .add = FALSE,
  .drop = dplyr::group_by_drop_default(self)
)

Arguments

...

variables or computations to group by. Computations are always done on the ungrouped data frame. To perform computations on the grouped data, you need to use a separate mutate() step before the group_by()

.add

When FALSE, the default, group_by() will override existing groups. To add to the existing groups, use .add = TRUE.

.drop

Drop groups formed by factor levels that don't appear in the data. The default is TRUE except when .data has been previously grouped with .drop = FALSE. See group_by_drop_default() for details.

Method slide()

Slides a given function over variables in an epi_archive object. See the documentation for the wrapper function epix_slide() for details. The parameter descriptions below are copied from there

Usage

epi_archive$slide(
  f,
  ...,
  before,
  ref_time_values,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_versions = FALSE
)

Arguments

f

Function, formula, or missing; together with ... specifies the computation to slide. To "slide" means to apply a computation over a sliding (a.k.a. "rolling") time window for each data group. The window is determined by the before parameter described below. One time step is typically one day or one week; see epi_slide details for more explanation. If a function, f must take an epi_df with the same column names as the archive's DT, minus the version column; followed by a one-row tibble containing the values of the grouping variables for the associated group; followed by a reference time value, usually as a Date object; followed by any number of named arguments. If a formula, f can operate directly on columns accessed via .x$var or .$var, as in ~ mean (.x$var) to compute a mean of a column var for each group-ref_time_value combination. The group key can be accessed via .y or .group_key, and the reference time value can be accessed via .z or .ref_time_value. If f is missing, then ... will specify the computation.

...

Additional arguments to pass to the function or formula specified via f. Alternatively, if f is missing, then ... is interpreted as an expression for tidy evaluation; in addition to referring to columns directly by name, the expression has access to .data and .env pronouns as in dplyr verbs, and can also refer to the .group_key and .ref_time_value. See details of epi_slide.

before

How far before each ref_time_value should the sliding window extend? If provided, should be a single, non-NA, integer-compatible number of time steps. This window endpoint is inclusive. For example, if before = 7, and one time step is one day, then to produce a value for a ref_time_value of January 8, we apply the given function or formula to data (for each group present) with time_values from January 1 onward, as they were reported on January 8. For typical disease surveillance sources, this will not include any data with a time_value of January 8, and, depending on the amount of reporting latency, may not include January 7 or even earlier time_values. (If instead the archive were to hold nowcasts instead of regular surveillance data, then we would indeed expect data for time_value January 8. If it were to hold forecasts, then we would expect data for time_values after January 8, and the sliding window would extend as far after each ref_time_value as needed to include all such time_values.)

ref_time_values

Reference time values / versions for sliding computations; each element of this vector serves both as the anchor point for the time_value window for the computation and the max_version as_of which we fetch data in this window. If missing, then this will set to a regularly-spaced sequence of values set to cover the range of versions in the DT plus the versions_end; the spacing of values will be guessed (using the GCD of the skips between values).

time_step

Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a positive integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

new_col_name

String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

as_list_col

Should the slide results be held in a list column, or be unchopped/unnested? Default is FALSE, in which case a list object returned by f would be unnested (using tidyr::unnest()), and, if the slide computations output data frames, the names of the resulting columns are given by prepending new_col_name to the names of the list elements.

names_sep

String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_versions

(Not the same as all_rows parameter of epi_slide.) If all_versions = TRUE, then f will be passed the version history (all version <= ref_time_value) for rows having time_value between ref_time_value - before and ref_time_value. Otherwise, f will be passed only the most recent version for every unique time_value. Default is FALSE.

Method clone()

The objects of this class are cloneable with this method.

Usage

epi_archive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

tib <- tibble::tibble(
  geo_value = rep(c("ca", "hi"), each = 5),
  time_value = rep(seq(as.Date("2020-01-01"),
    by = 1, length.out = 5
  ), times = 2),
  version = rep(seq(as.Date("2020-01-02"),
    by = 1, length.out = 5
  ), times = 2),
  value = rnorm(10, mean = 2, sd = 1)
)

toy_epi_archive <- tib %>% epi_archive$new(
  geo_type = "state",
  time_type = "day"
)
toy_epi_archive
#> → An `epi_archive` object, with metadata:
#> ℹ Min/max time values: 2020-01-01 / 2020-01-05
#> ℹ First/last version with update: 2020-01-02 / 2020-01-06
#> ℹ Versions end: 2020-01-06
#> ℹ Public R6 methods: initialize, print, as_of, fill_through_version,
#>   truncate_versions_after, merge, group_by, slide, and clone
#> ℹ A preview of the table (10 rows x 4 columns):
#> Key: <geo_value, time_value, version>
#>     geo_value time_value    version      value
#>        <char>     <Date>     <Date>      <num>
#>  1:        ca 2020-01-01 2020-01-02 1.17321105
#>  2:        ca 2020-01-02 2020-01-03 0.48760035
#>  3:        ca 2020-01-03 2020-01-04 2.93536319
#>  4:        ca 2020-01-04 2020-01-05 2.17648861
#>  5:        ca 2020-01-05 2020-01-06 2.24368546
#>  6:        hi 2020-01-01 2020-01-02 3.62354888
#>  7:        hi 2020-01-02 2020-01-03 2.11203808
#>  8:        hi 2020-01-03 2020-01-04 1.86600299
#>  9:        hi 2020-01-04 2020-01-05 0.08991253
#> 10:        hi 2020-01-05 2020-01-06 1.72076276