epi_df object

An epi_df is a tibble with certain minimal column structure and metadata. It can be seen as a snapshot of a data set that contains the most up-to-date values of some signal variables of interest, as of a given time.

Details

An epi_df is a tibble with (at least) the following columns:

geo_value: the geographic value associated with each row of measurements.

time_value: the time value associated with each row of measurements.

Other columns can be considered as measured variables, which we also refer to as signal variables. An epi_df object also has metadata with (at least) the following fields:

geo_type: the type for the geo values.

time_type: the type for the time values.

as_of: the time value at which the given data were available.

Metadata for an epi_df object x can be accessed (and altered) via attributes(x)$metadata. The first two fields in the above list, geo_type and time_type, can usually be inferred from the geo_value and time_value columns, respectively. They are not currently used by any downstream functions in the epiprocess package, and serve only as useful bits of information to convey about the data set at hand. More information on their coding is given below.

The last field in the above list, as_of, is one of the most unique aspects of an epi_df object. In brief, we can think of an epi_df object as a single snapshot of a data set that contains the most up-to-date values of the signals variables, as of the time specified in the as_of field.

A companion object is the epi_archive object, which contains the full version history of a given data set. Revisions are common in many types of epidemiological data streams, and paying attention to data revisions can be important for all sorts of downstream data analysis and modeling tasks. See the documentation for epi_archive for more details on how data versioning works in the epiprocess package (including how to generate epi_df objects, as data snapshots, from an epi_archive object).

Geo Types

The following geo types are recognized in an epi_df.

"county": each observation corresponds to a U.S. county; coded by 5-digit FIPS code.

"hrr": each observation corresponds to a U.S. hospital referral region (designed to represent regional healthcare markets); there are 306 HRRs in the U.S; coded by number (nonconsecutive, between 1 and 457).

"state": each observation corresponds to a U.S. state; coded by 2-digit postal abbreviation (lowercase); note that Puerto Rico is "pr" and Washington D.C. is "dc".

"hhs": each observation corresponds to a U.S. HHS region; coded by number (consecutive, between 1 and 10).

"nation": each observation corresponds to a country; coded by ISO 31661- alpha-2 country codes (lowercase).

An unrecognizable geo type is labeled "custom".

Time Types

The following time types are recognized in an epi_df.

"day-time": each observation corresponds to a time on a given day (measured to the second); coded as a POSIXct object, as in as.POSIXct("2022-01-31 18:45:40").

"day": each observation corresponds to a day; coded as a Date object, as in as.Date("2022-01-31").

"week": each observation corresponds to a week; the alignment can be arbitrary (as to whether a week starts on a Monday, Tuesday); coded as a Date object, representing the start date of week.

"yearweek": each observation corresponds to a week; the alignment can be arbitrary; coded as a tsibble::yearweek object, where the alignment is stored in the week_start field of its attributes.

"yearmonth": each observation corresponds to a month; coded as a tsibble::yearmonth object.

"yearquarter": each observation corresponds to a quarter; coded as a tsibble::yearquarter object.

"year": each observation corresponds to a year; coded as an integer greater than or equal to 1582.

An unrecognizable time type is labeled "custom".