Slide a function over variables in an epi_tibble object

Slides a given function over variables in an epi_tibble object. See the slide vignette for examples.

epi_slide(
  x,
  slide_fun,
  n = 14,
  align = c("right", "center", "left"),
  before,
  new_col_name = "slide_value",
  new_col_type = c("dbl", "int", "lgl", "chr", "list"),
  time_step,
  ...
)

Arguments

x
The epi_tibble object under consideration.

slide_fun
Function or formula to slide over variables in x. To "slide" means to apply a function or formula over a running window of n time steps (where one time step is typically one day or one week; see details for more explanation). If a function, slide_fun must take x, a data frame with the same column names as the original object; followed by any number of named arguments; and ending with .... If a formula, slide_fun can operate directly on columns accessed via .x$var, as in ~ mean(.x$var) to compute a mean of a column var over a sliding window of n time steps.

n
Number of time steps to use in the running window. For example, if n = 5, one time step is one day, and the alignment is "right", then to produce a value on November 5, we apply the given function or formula to data in between November 1 and 5. Default is 14.

align
One of "right", "center", or "left", indicating the alignment of the sliding window relative to the reference time point. If the alignment is "center" and n is even, then one more time point will be used before the reference time point than after. Default is "right".

before
Optional integer, in the range to 0 to n-1 (inclusive), specifying the number of time points to use in the sliding window strictly before the reference time point. For example, setting before = n-1 would be the same as setting align = "right". The current argument allows for more flexible specification of alignment than the align parameter, and if specified, then it overrides align.

new_col_name
String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

new_col_type
One of "dbl", "int", "lgl", "chr", or "list", indicating the data type (tibble abbreviation) for the new column. Default is "dbl".

time_step
Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the metadata. This function must take a positive integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct, that is, if time_type is "day-time").

...
Additional arguments to pass to the function or formula specified via slide_fun.

Value

An epi_tibble object given by appending a new column to x, named according to the new_col_name argument, containing the slide values.

Details

To "slide" means to apply a function or formula over a running window of n time steps, where the unit (the meaning of one time step) is determined by the time_type field in the metadata: the unit is one day if time_type is either "day-time" or "day", and the unit is one week if time_type is "week". The time step can also be set explicitly using the time_step argument (which if specified would override the default choice based on the metadata).

In order to slide a function or formula in a suitable grouped fashion (for example, per geo value), we can use group_by() before the call to epi_slide().