Slide a function over variables in an epi_df object

Slides a given function over variables in an epi_df object. See the slide vignette for examples.

Usage

epi_slide(
  x,
  f,
  ...,
  n = 7,
  ref_time_values,
  align = c("right", "center", "left"),
  before,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_rows = FALSE
)

Arguments

x

The epi_df object under consideration.

f

Function or formula to slide over variables in x. To "slide" means to apply a function or formula over a running window of n time steps (where one time step is typically one day or one week; see details for more explanation). If a function, f must take x, a data frame with the same column names as the original object; followed by any number of named arguments; and ending with .... If a formula, f can operate directly on columns accessed via .x$var, as in ~ mean(.x$var) to compute a mean of a column var over a sliding window of n time steps.

...

Additional arguments to pass to the function or formula specified via f. Alternatively, if f is missing, then the current argument is interpreted as an expression for tidy evaluation. See details.

n

Number of time steps to use in the running window. For example, if n = 7, one time step is one day, and the alignment is "right", then to produce a value on January 7 we apply the given function or formula to data in between January 1 and 7. Default is 7.

ref_time_values

Time values for sliding computations, meaning, each element of this vector serves as the reference time point for one sliding window. If missing, then this will be set to all unique time values in the underlying data table, by default.

align

One of "right", "center", or "left", indicating the alignment of the sliding window relative to the reference time point. If the alignment is "center" and n is even, then one more time point will be used after the reference time point than before. Default is "right".

before

Positive integer less than n, specifying the number of time points to use in the sliding window strictly before the reference time point. For example, setting before = n-1 would be the same as setting align = "right". The before argument allows for more flexible specification of alignment than the align parameter, and if specified, overrides align.

time_step

Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a positive integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

new_col_name

String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

as_list_col

Should the new column be stored as a list column? Default is FALSE, in which case a list object returned by f would be unnested (using tidyr::unnest()), and the names of the resulting columns are given by prepending new_col_name to the names of the list elements.

names_sep

String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_rows

If all_rows = TRUE, then all rows of x will be kept in the output; otherwise, there will be one row for each time value in x that acts as a reference time value. Default is FALSE.

Value

An epi_df object given by appending a new column to x, named according to the new_col_name argument.

Details

To "slide" means to apply a function or formula over a running window of n time steps, where the unit (the meaning of one time step) is implicitly defined by the way the time_value column treats addition and subtraction; for example, if the time values are coded as Date objects, then one time step is one day, since as.Date("2022-01-01") + 1 equals as.Date("2022-01-02"). Alternatively, the time step can be set explicitly using the time_step argument (which if specified would override the default choice based on time_value column). If less than n time steps are available at any given reference time value, then epi_slide() still attempts to perform the computation anyway (it does not require a complete window). The issue of what to do with partial computations (those run on incomplete windows) is therefore left up to the user, either through the specified function or formula f, or through post-processing.

If f is missing, then an expression for tidy evaluation can be specified, for example, as in:

epi_slide(x, cases_7dav = mean(cases), n = 7)

which would be equivalent to:

epi_slide(x, function(x, ...) mean(x$cases), n = 7,
          new_col_name = "cases_7dav")

Thus, to be clear, when the computation is specified via an expression for tidy evaluation (first example, above), then the name for the new column is inferred from the given expression and overrides any name passed explicitly through the new_col_name argument.

When f is a named function with arguments, if a tibble with an unnamed grouping variable is passed in as the method argument to f, include a parameter for the grouping-variable in function() just prior to specifying the method to prevent that from being overridden. For example:

# Construct an tibble with an unnamed grouping variable
edf = bind_rows(tibble(geo_value = "ak", time_value = as.Date("2020-01-01") 
          + 1:10, x1=1:10, y=1:10 + rnorm(10L))) %>% 
  as_epi_df()

# Now, include a row parameter for the grouping variable in the tibble, 
# which we denote as g, just prior to method = "qr"
# Note that if g was not included below, then the method = "qr" would be 
# overridden, as described above
edf %>%
group_by(geo_value) %>%
epi_slide(function(x, g, method="qr", ...) tibble(model=list(
          lm(y ~ x1, x, method=method))), n=7L) 

Examples

 # slide a 7-day trailing average formula on cases
  jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(cases_7dav = mean(cases), n = 7, 
            align = "right") %>% 
  # rmv a nonessential var. to ensure new col is printed
  dplyr::select(-death_rate_7d_av) 
#> # A tibble: 4,026 × 6
#> # Groups:   geo_value [6]
#>    geo_value time_value case_rate_7d_av cases cases_7d_av cases_7dav
#>    <chr>     <date>               <dbl> <dbl>       <dbl>      <dbl>
#>  1 ca        2020-03-01         0.00327     6        1.29       6   
#>  2 ca        2020-03-02         0.00435     4        1.71       5   
#>  3 ca        2020-03-03         0.00617     6        2.43       5.33
#>  4 ca        2020-03-04         0.00980    11        3.86       6.75
#>  5 ca        2020-03-05         0.0134     10        5.29       7.4 
#>  6 ca        2020-03-06         0.0200     18        7.86       9.17
#>  7 ca        2020-03-07         0.0294     26       11.6       11.6 
#>  8 ca        2020-03-08         0.0341     19       13.4       13.4 
#>  9 ca        2020-03-09         0.0410     23       16.1       16.1 
#> 10 ca        2020-03-10         0.0468     22       18.4       18.4 
#> # … with 4,016 more rows
 
 # slide a left-aligned 7-day average
  jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(cases_7dav = mean(cases), n = 7, 
            align = "left") %>% 
  # rmv a nonessential var. to ensure new col is printed
  dplyr::select(-death_rate_7d_av) 
#> # A tibble: 4,026 × 6
#> # Groups:   geo_value [6]
#>    geo_value time_value case_rate_7d_av cases cases_7d_av cases_7dav
#>    <chr>     <date>               <dbl> <dbl>       <dbl>      <dbl>
#>  1 ca        2020-03-01         0.00327     6        1.29       11.6
#>  2 ca        2020-03-02         0.00435     4        1.71       13.4
#>  3 ca        2020-03-03         0.00617     6        2.43       16.1
#>  4 ca        2020-03-04         0.00980    11        3.86       18.4
#>  5 ca        2020-03-05         0.0134     10        5.29       20.4
#>  6 ca        2020-03-06         0.0200     18        7.86       25.1
#>  7 ca        2020-03-07         0.0294     26       11.6        30.1
#>  8 ca        2020-03-08         0.0341     19       13.4        34.4
#>  9 ca        2020-03-09         0.0410     23       16.1        37.3
#> 10 ca        2020-03-10         0.0468     22       18.4        56.7
#> # … with 4,016 more rows
 
 # nested new columns
 jhu_csse_daily_subset %>% 
 group_by(geo_value) %>%
 epi_slide(a = data.frame(cases_2dav = mean(cases), 
                          cases_2dma = mad(cases)),
           n = 2, as_list_col = TRUE)
#> An `epi_df` object, 4,026 x 7 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2022-05-23 20:17:07
#> 
#> # A tibble: 4,026 × 7
#> # Groups:   geo_value [6]
#>    geo_value time_value case_rate_7d_av death_rate_7d_av cases cases_7d_av a    
#>  * <chr>     <date>               <dbl>            <dbl> <dbl>       <dbl> <lis>
#>  1 ca        2020-03-01         0.00327         0            6        1.29 <df> 
#>  2 ca        2020-03-02         0.00435         0            4        1.71 <df> 
#>  3 ca        2020-03-03         0.00617         0            6        2.43 <df> 
#>  4 ca        2020-03-04         0.00980         0.000363    11        3.86 <df> 
#>  5 ca        2020-03-05         0.0134          0.000363    10        5.29 <df> 
#>  6 ca        2020-03-06         0.0200          0.000363    18        7.86 <df> 
#>  7 ca        2020-03-07         0.0294          0.000363    26       11.6  <df> 
#>  8 ca        2020-03-08         0.0341          0.000363    19       13.4  <df> 
#>  9 ca        2020-03-09         0.0410          0.000726    23       16.1  <df> 
#> 10 ca        2020-03-10         0.0468          0.000726    22       18.4  <df> 
#> # … with 4,016 more rows