Slide a function over variables in an epi_df object

Slides a given function over variables in an epi_df object. See the slide vignette for examples.

epi_slide(
  x,
  f,
  ...,
  n = 7,
  ref_time_values,
  align = c("right", "center", "left"),
  before,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_rows = FALSE
)

Arguments

x

The epi_df object under consideration.

f

Function or formula to slide over variables in x. To "slide" means to apply a function or formula over a running window of n time steps (where one time step is typically one day or one week; see details for more explanation). If a function, f must take x, a data frame with the same column names as the original object; followed by any number of named arguments; and ending with .... If a formula, f can operate directly on columns accessed via .x$var, as in ~ mean(.x$var) to compute a mean of a column var over a sliding window of n time steps.

...

Additional arguments to pass to the function or formula specified via f. Alternatively, if f is missing, then the current argument is interpreted as an expression for tidy evaluation. See details.

n

Number of time steps to use in the running window. For example, if n = 7, one time step is one day, and the alignment is "right", then to produce a value on January 7 we apply the given function or formula to data in between January 1 and 7. Default is 7.

ref_time_values

Time values for sliding computations, meaning, each element of this vector serves as the reference time point for one sliding window. If missing, then this will be set to all unique time values in the underlying data table, by default.

align

One of "right", "center", or "left", indicating the alignment of the sliding window relative to the reference time point. If the alignment is "center" and n is even, then one more time point will be used after the reference time point than before. Default is "right".

before

Positive integer less than n, specifying the number of time points to use in the sliding window strictly before the reference time point. For example, setting before = n-1 would be the same as setting align = "right". The before argument allows for more flexible specification of alignment than the align parameter, and if specified, overrides align.

time_step

Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a positive integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

new_col_name

String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

as_list_col

Should the new column be stored as a list column? Default is FALSE, in which case a list object returned by f would be unnested (using tidyr::unnest()), and the names of the resulting columns are given by prepending new_col_name to the names of the list elements.

names_sep

String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_rows

If all_rows = TRUE, then all rows of x will be kept in the output; otherwise, there will be one row for each time value in x that acts as a reference time value. Default is FALSE.

Value

An epi_df object given by appending a new column to x, named according to the new_col_name argument.

Details

To "slide" means to apply a function or formula over a running window of n time steps, where the unit (the meaning of one time step) is implicitly defined by the way the time_value column treats addition and subtraction; for example, if the time values are coded as Date objects, then one time step is one day, since as.Date("2022-01-01") + 1 equals as.Date("2022-01-02"). Alternatively, the time step can be set explicitly using the time_step argument (which if specified would override the default choice based on time_value column). If less than n time steps are available at any given reference time value, then epi_slide() still attempts to perform the computation anyway (it does not require a complete window). The issue of what to do with partial computations (those run on incomplete windows) is therefore left up to the user, either through the specified function or formula f, or through post-processing.

If f is missing, then an expression for tidy evaluation can be specified, for example, as in:

epi_slide(x, cases_7dav = mean(cases), n = 7)

which would be equivalent to:

epi_slide(x, function(x, ...) mean(x$cases), n = 7,
          new_col_name = "cases_7dav")

Thus, to be clear, when the computation is specified via an expression for tidy evaluation (first example, above), then the name for the new column is inferred from the given expression and overrides any name passed explicitly through the new_col_name argument.