Slide a function over variables in an epi_df object

Slides a given function over variables in an epi_df object. See the slide vignette for examples.

Usage

epi_slide(
  x,
  f,
  ...,
  before,
  after,
  ref_time_values,
  time_step,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_rows = FALSE
)

Arguments

x

The epi_df object under consideration, grouped or ungrouped. If ungrouped, all data in x will be treated as part of a single data group.

f

Function, formula, or missing; together with ... specifies the computation to slide. To "slide" means to apply a computation within a sliding (a.k.a. "rolling") time window for each data group. The window is determined by the before and after parameters described below. One time step is typically one day or one week; see details for more explanation. If a function, f must take a data frame with the same column names as the original object, minus any grouping variables, containing the time window data for one group-ref_time_value combination; followed by a one-row tibble containing the values of the grouping variables for the associated group; followed by any number of named arguments. If a formula, f can operate directly on columns accessed via .x$var or .$var, as in ~mean(.x$var) to compute a mean of a column var for each ref_time_value-group combination. The group key can be accessed via .y. If f is missing, then ... will specify the computation.

...

Additional arguments to pass to the function or formula specified via f. Alternatively, if f is missing, then the ... is interpreted as an expression for tidy evaluation; in addition to referring to columns directly by name, the expression has access to .data and .env pronouns as in dplyr verbs, and can also refer to .x, .group_key, and .ref_time_value. See details.

before, after

How far before and after each ref_time_value should the sliding window extend? At least one of these two arguments must be provided; the other's default will be 0. Any value provided for either argument must be a single, non-NA, non-negative, integer-compatible number of time steps. Endpoints of the window are inclusive. Common settings:

For trailing/right-aligned windows from ref_time_value - time_step (k) to ref_time_value: either pass before=k by itself, or pass before=k, after=0.

For center-aligned windows from ref_time_value - time_step(k) to ref_time_value + time_step(k): pass before=k, after=k.

For leading/left-aligned windows from ref_time_value to ref_time_value + time_step(k): either pass pass after=k by itself, or pass before=0, after=k. See "Details:" about the definition of a time step,(non)treatment of missing rows within the window, and avoiding warnings about before&after settings for a certain uncommon use case.

ref_time_values

Time values for sliding computations, meaning, each element of this vector serves as the reference time point for one sliding window. If missing, then this will be set to all unique time values in the underlying data table, by default.

time_step

Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a non-negative integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

new_col_name

String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

as_list_col

Should the slide results be held in a list column, or be unchopped/unnested? Default is FALSE, in which case a list object returned by f would be unnested (using tidyr::unnest()), and, if the slide computations output data frames, the names of the resulting columns are given by prepending new_col_name to the names of the list elements.

names_sep

String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_rows

If all_rows = TRUE, then all rows of x will be kept in the output even with ref_time_values provided, with some type of missing value marker for the slide computation output column(s) for time_values outside ref_time_values; otherwise, there will be one row for each row in x that had a time_value in ref_time_values. Default is FALSE. The missing value marker is the result of vctrs::vec_casting NA to the type of the slide computation output. If using as_list_col = TRUE, note that the missing marker is a NULL entry in the list column; for certain operations, you might want to replace these NULL entries with a different NA marker.

Value

An epi_df object given by appending one or more new columns to x, named according to the new_col_name argument.

Details

To "slide" means to apply a function or formula over a rolling window of time steps for each data group, where the window is centered at a reference time and left and right endpoints are given by the before and after arguments. The unit (the meaning of one time step) is implicitly defined by the way the time_value column treats addition and subtraction; for example, if the time values are coded as Date objects, then one time step is one day, since as.Date("2022-01-01") + 1 equals as.Date("2022-01-02"). Alternatively, the time step can be set explicitly using the time_step argument (which if specified would override the default choice based on time_value column). If there are not enough time steps available to complete the window at any given reference time, then epi_slide() still attempts to perform the computation anyway (it does not require a complete window). The issue of what to do with partial computations (those run on incomplete windows) is therefore left up to the user, either through the specified function or formula f, or through post-processing. For a centrally-aligned slide of n time_values in a sliding window, set before = (n-1)/2 and after = (n-1)/2 when the number of time_values in a sliding window is odd and before = n/2-1 and after = n/2 when n is even.

Sometimes, we want to experiment with various trailing or leading window widths and compare the slide outputs. In the (uncommon) case where zero-width windows are considered, manually pass both the before and after arguments in order to prevent potential warnings. (E.g., before=k with k=0 and after missing may produce a warning. To avoid warnings, use before=k, after=0 instead; otherwise, it looks too much like a leading window was intended, but the after argument was forgotten or misspelled.)

If f is missing, then an expression for tidy evaluation can be specified, for example, as in:

epi_slide(x, cases_7dav = mean(cases), before = 6)

which would be equivalent to:

epi_slide(x, function(x, g) mean(x$cases), before = 6,
          new_col_name = "cases_7dav")

Thus, to be clear, when the computation is specified via an expression for tidy evaluation (first example, above), then the name for the new column is inferred from the given expression and overrides any name passed explicitly through the new_col_name argument.

See also

Examples

# slide a 7-day trailing average formula on cases
# Simple sliding means and sums are much faster to do using
# the `epi_slide_mean` and `epi_slide_sum` functions instead.
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(cases_7dav = mean(cases), before = 6) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_7dav) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6       6   
#>  2 ca        2020-03-02     4       5   
#>  3 ca        2020-03-03     6       5.33
#>  4 ca        2020-03-04    11       6.75
#>  5 ca        2020-03-05    10       7.4 
#>  6 ca        2020-03-06    18       9.17
#>  7 ca        2020-03-07    26      11.6 
#>  8 ca        2020-03-08    19      13.4 
#>  9 ca        2020-03-09    23      16.1 
#> 10 ca        2020-03-10    22      18.4 
#> # ℹ 4,016 more rows

# slide a 7-day leading average
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(cases_7dav = mean(cases), after = 6) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_7dav) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6       11.6
#>  2 ca        2020-03-02     4       13.4
#>  3 ca        2020-03-03     6       16.1
#>  4 ca        2020-03-04    11       18.4
#>  5 ca        2020-03-05    10       20.4
#>  6 ca        2020-03-06    18       25.1
#>  7 ca        2020-03-07    26       30.1
#>  8 ca        2020-03-08    19       34.4
#>  9 ca        2020-03-09    23       37.3
#> 10 ca        2020-03-10    22       56.7
#> # ℹ 4,016 more rows

# slide a 7-day centre-aligned average
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(cases_7dav = mean(cases), before = 3, after = 3) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_7dav) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6       6.75
#>  2 ca        2020-03-02     4       7.4 
#>  3 ca        2020-03-03     6       9.17
#>  4 ca        2020-03-04    11      11.6 
#>  5 ca        2020-03-05    10      13.4 
#>  6 ca        2020-03-06    18      16.1 
#>  7 ca        2020-03-07    26      18.4 
#>  8 ca        2020-03-08    19      20.4 
#>  9 ca        2020-03-09    23      25.1 
#> 10 ca        2020-03-10    22      30.1 
#> # ℹ 4,016 more rows

# slide a 14-day centre-aligned average
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(cases_14dav = mean(cases), before = 6, after = 7) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_14dav) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_14dav
#>  * <chr>     <date>     <dbl>       <dbl>
#>  1 ca        2020-03-01     6        12.5
#>  2 ca        2020-03-02     4        13.7
#>  3 ca        2020-03-03     6        14.5
#>  4 ca        2020-03-04    11        15.5
#>  5 ca        2020-03-05    10        17.8
#>  6 ca        2020-03-06    18        20.5
#>  7 ca        2020-03-07    26        23  
#>  8 ca        2020-03-08    19        25.4
#>  9 ca        2020-03-09    23        36.4
#> 10 ca        2020-03-10    22        42  
#> # ℹ 4,016 more rows

# nested new columns
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide(
    a = data.frame(
      cases_2dav = mean(cases),
      cases_2dma = mad(cases)
    ),
    before = 1, as_list_col = TRUE
  ) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 7 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 7
#>    geo_value time_value cases cases_7d_av case_rate_7d_av death_rate_7d_av a    
#>  * <chr>     <date>     <dbl>       <dbl>           <dbl>            <dbl> <lis>
#>  1 ca        2020-03-01     6        1.29         0.00327         0        <df> 
#>  2 ca        2020-03-02     4        1.71         0.00435         0        <df> 
#>  3 ca        2020-03-03     6        2.43         0.00617         0        <df> 
#>  4 ca        2020-03-04    11        3.86         0.00980         0.000363 <df> 
#>  5 ca        2020-03-05    10        5.29         0.0134          0.000363 <df> 
#>  6 ca        2020-03-06    18        7.86         0.0200          0.000363 <df> 
#>  7 ca        2020-03-07    26       11.6          0.0294          0.000363 <df> 
#>  8 ca        2020-03-08    19       13.4          0.0341          0.000363 <df> 
#>  9 ca        2020-03-09    23       16.1          0.0410          0.000726 <df> 
#> 10 ca        2020-03-10    22       18.4          0.0468          0.000726 <df> 
#> # ℹ 4,016 more rows