Optimized slide function for performing common rolling computations on an epi_df object

Slides an n-timestep data.table::froll or slider::summary-slide function over variables in an epi_df object. See the slide vignette for examples.

Usage

epi_slide_opt(
  x,
  col_names,
  f,
  ...,
  before = NULL,
  after = NULL,
  ref_time_values = NULL,
  new_col_name = NULL,
  as_list_col = NULL,
  names_sep = NULL,
  all_rows = FALSE
)

Arguments

x

The epi_df object under consideration, grouped or ungrouped. If ungrouped, all data in x will be treated as part of a single data group.

col_names

<tidy-select> An unquoted column name(e.g., cases), multiple column names (e.g., c(cases, deaths)), or other tidy-select expression. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables. If you have the desired column names stored in a vector vars, use col_names = all_of(vars).

The tidy-selection renaming interface is not supported, and cannot be used to provide output column names; if you want to customize the output column names, use dplyr::rename after the slide.

f

Function; together with ... specifies the computation to slide. f must be one of data.table's rolling functions (frollmean, frollsum, frollapply. See data.table::roll) or one of slider's specialized sliding functions (slide_mean, slide_sum, etc. See slider::summary-slide). To "slide" means to apply a computation within a sliding (a.k.a. "rolling") time window for each data group. The window is determined by the before and after parameters described below. One time step is typically one day or one week; see details for more explanation.

The optimized data.table and slider functions can't be directly passed as the computation function in epi_slide without careful handling to make sure each computation group is made up of the n dates rather than n points. epi_slide_opt (and wrapper functions epi_slide_mean and epi_slide_sum) take care of window completion automatically to prevent associated errors.

...

Additional arguments to pass to the slide computation f, for example, na.rm and algo if f is a data.table function. If f is a data.table function, it is automatically passed the data x to operate on, the window size n, and the alignment align. Providing these args via ... will cause an error. If f is a slider function, it is automatically passed the data x to operate on, and number of points before and after to use in the computation.

before, after

How far before and after each ref_time_value should the sliding window extend? At least one of these two arguments must be provided; the other's default will be 0. The accepted values for these depend on the type of the time_value column:

if it is a Date and the cadence is daily, then they can be integers (which will be interpreted in units of days) or difftimes with units "days"

if it is a Date and the cadence is weekly, then they must be difftimes with units "weeks"

if it is an integer, then they must be integers

Endpoints of the window are inclusive. Common settings:

For trailing/right-aligned windows from ref_time_value - k to ref_time_value: either pass before=k by itself, or pass before=k, after=0.

For center-aligned windows from ref_time_value - k to ref_time_value + k: pass before=k, after=k.

For leading/left-aligned windows from ref_time_value to ref_time_value + k: either pass pass after=k by itself, or pass before=0, after=k.

See "Details:" on how missing rows are handled within the window.

ref_time_values

Time values for sliding computations, meaning, each element of this vector serves as the reference time point for one sliding window. If missing, then this will be set to all unique time values in the underlying data table, by default.

new_col_name

Character vector indicating the name(s) of the new column(s) that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to any existing column names will overwrite those columns. If names_sep is NULL, new_col_name must be the same length as col_names.

as_list_col

Not supported. Included to match epi_slide interface.

names_sep

String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_rows

If all_rows = TRUE, then all rows of x will be kept in the output even with ref_time_values provided, with some type of missing value marker for the slide computation output column(s) for time_values outside ref_time_values; otherwise, there will be one row for each row in x that had a time_value in ref_time_values. Default is FALSE. The missing value marker is the result of vctrs::vec_casting NA to the type of the slide computation output. If using as_list_col = TRUE, note that the missing marker is a NULL entry in the list column; for certain operations, you might want to replace these NULL entries with a different NA marker.

Value

An epi_df object given by appending one or more new columns to x, named according to the new_col_name argument.

Details

To "slide" means to apply a function over a rolling window of time steps for each data group, where the window is centered at a reference time and left and right endpoints are given by the before and after arguments. If there are not enough time steps available to complete the window at any given reference time, then epi_slide_*() will fail; it requires a complete window to perform the computation. For a centrally-aligned slide of n time_values in a sliding window, set before = (n-1)/2 and after = (n-1)/2 when the number of time_values in a sliding window is odd and before = n/2-1 and after = n/2 when n is even.

Sometimes, we want to experiment with various trailing or leading window widths and compare the slide outputs. In the (uncommon) case where zero-width windows are considered, manually pass both the before and after arguments.

See also

Examples

# slide a 7-day trailing average formula on cases. This can also be done with `epi_slide_mean`
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide_opt(
    cases,
    f = data.table::frollmean, before = 6
  ) %>%
  # Remove a nonessential var. to ensure new col is printed, and rename new col
  dplyr::select(geo_value, time_value, cases, cases_7dav = slide_value_cases) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6       NA  
#>  2 ca        2020-03-02     4       NA  
#>  3 ca        2020-03-03     6       NA  
#>  4 ca        2020-03-04    11       NA  
#>  5 ca        2020-03-05    10       NA  
#>  6 ca        2020-03-06    18       NA  
#>  7 ca        2020-03-07    26       11.6
#>  8 ca        2020-03-08    19       13.4
#>  9 ca        2020-03-09    23       16.1
#> 10 ca        2020-03-10    22       18.4
#> # ℹ 4,016 more rows

# slide a 7-day trailing average formula on cases. Adjust `frollmean` settings for speed
# and accuracy, and to allow partially-missing windows.
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide_opt(
    cases,
    f = data.table::frollmean, before = 6,
    # `frollmean` options
    na.rm = TRUE, algo = "exact", hasNA = TRUE
  ) %>%
  dplyr::select(geo_value, time_value, cases, cases_7dav = slide_value_cases) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6       6   
#>  2 ca        2020-03-02     4       5   
#>  3 ca        2020-03-03     6       5.33
#>  4 ca        2020-03-04    11       6.75
#>  5 ca        2020-03-05    10       7.4 
#>  6 ca        2020-03-06    18       9.17
#>  7 ca        2020-03-07    26      11.6 
#>  8 ca        2020-03-08    19      13.4 
#>  9 ca        2020-03-09    23      16.1 
#> 10 ca        2020-03-10    22      18.4 
#> # ℹ 4,016 more rows

# slide a 7-day leading average
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide_opt(
    cases,
    f = slider::slide_mean, after = 6
  ) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_7dav = slide_value_cases) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6       11.6
#>  2 ca        2020-03-02     4       13.4
#>  3 ca        2020-03-03     6       16.1
#>  4 ca        2020-03-04    11       18.4
#>  5 ca        2020-03-05    10       20.4
#>  6 ca        2020-03-06    18       25.1
#>  7 ca        2020-03-07    26       30.1
#>  8 ca        2020-03-08    19       34.4
#>  9 ca        2020-03-09    23       37.3
#> 10 ca        2020-03-10    22       56.7
#> # ℹ 4,016 more rows

# slide a 7-day centre-aligned sum. This can also be done with `epi_slide_sum`
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide_opt(
    cases,
    f = data.table::frollsum, before = 3, after = 3
  ) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_7dav = slide_value_cases) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dav
#>  * <chr>     <date>     <dbl>      <dbl>
#>  1 ca        2020-03-01     6         NA
#>  2 ca        2020-03-02     4         NA
#>  3 ca        2020-03-03     6         NA
#>  4 ca        2020-03-04    11         81
#>  5 ca        2020-03-05    10         94
#>  6 ca        2020-03-06    18        113
#>  7 ca        2020-03-07    26        129
#>  8 ca        2020-03-08    19        143
#>  9 ca        2020-03-09    23        176
#> 10 ca        2020-03-10    22        211
#> # ℹ 4,016 more rows