Optimized slide function for performing rolling sums on an epi_df object

Slides an n-timestep sum over variables in an epi_df object. See the slide vignette for examples.

Usage

epi_slide_sum(
  x,
  col_names,
  ...,
  before,
  after,
  ref_time_values,
  time_step,
  new_col_name = "slide_value",
  as_list_col = NULL,
  names_sep = "_",
  all_rows = FALSE
)

Arguments

x

The epi_df object under consideration, grouped or ungrouped. If ungrouped, all data in x will be treated as part of a single data group.

col_names

A single tidyselection or a tidyselection vector of the names of one or more columns for which to calculate the rolling mean.

...

Additional arguments to pass to data.table::frollsum, for example, na.rm and algo. data.table::frollsum is automatically passed the data x to operate on, the window size n, and the alignment align. Providing these args via ... will cause an error.

before, after

How far before and after each ref_time_value should the sliding window extend? At least one of these two arguments must be provided; the other's default will be 0. Any value provided for either argument must be a single, non-NA, non-negative, integer-compatible number of time steps. Endpoints of the window are inclusive. Common settings:

For trailing/right-aligned windows from ref_time_value - time_step (k) to ref_time_value: either pass before=k by itself, or pass before=k, after=0.

For center-aligned windows from ref_time_value - time_step(k) to ref_time_value + time_step(k): pass before=k, after=k.

For leading/left-aligned windows from ref_time_value to ref_time_value + time_step(k): either pass pass after=k by itself, or pass before=0, after=k. See "Details:" about the definition of a time step,(non)treatment of missing rows within the window, and avoiding warnings about before&after settings for a certain uncommon use case.

ref_time_values

Time values for sliding computations, meaning, each element of this vector serves as the reference time point for one sliding window. If missing, then this will be set to all unique time values in the underlying data table, by default.

time_step

Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a non-negative integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

new_col_name

Character vector indicating the name(s) of the new column(s) that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to any existing column names will overwrite those columns. If names_sep is NULL, new_col_name must be the same length as col_names.

as_list_col

Not supported. Included to match epi_slide interface.

names_sep

String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_rows

If all_rows = TRUE, then all rows of x will be kept in the output even with ref_time_values provided, with some type of missing value marker for the slide computation output column(s) for time_values outside ref_time_values; otherwise, there will be one row for each row in x that had a time_value in ref_time_values. Default is FALSE. The missing value marker is the result of vctrs::vec_casting NA to the type of the slide computation output. If using as_list_col = TRUE, note that the missing marker is a NULL entry in the list column; for certain operations, you might want to replace these NULL entries with a different NA marker.

Value

An epi_df object given by appending one or more new columns to x, named according to the new_col_name argument.

Details

Wrapper around epi_slide_opt with f = datatable::frollsum.

To "slide" means to apply a function over a rolling window of time steps for each data group, where the window is centered at a reference time and left and right endpoints are given by the before and after arguments. The unit (the meaning of one time step) is implicitly defined by the way the time_value column treats addition and subtraction; for example, if the time values are coded as Date objects, then one time step is one day, since as.Date("2022-01-01") + 1 equals as.Date ("2022-01-02"). Alternatively, the time step can be set explicitly using the time_step argument (which if specified would override the default choice based on time_value column). If there are not enough time steps available to complete the window at any given reference time, then epi_slide_*() will fail; it requires a complete window to perform the computation. For a centrally-aligned slide of n time_values in a sliding window, set before = (n-1)/2 and after = (n-1)/2 when the number of time_values in a sliding window is odd and before = n/2-1 and after = n/2 when n is even.

Sometimes, we want to experiment with various trailing or leading window widths and compare the slide outputs. In the (uncommon) case where zero-width windows are considered, manually pass both the before and after arguments in order to prevent potential warnings. (E.g., before=k with k=0 and after missing may produce a warning. To avoid warnings, use before=k, after=0 instead; otherwise, it looks too much like a leading window was intended, but the after argument was forgotten or misspelled.)

See also

Examples

# slide a 7-day trailing sum formula on cases
jhu_csse_daily_subset %>%
  group_by(geo_value) %>%
  epi_slide_sum(cases, new_col_name = "cases_7dsum", names_sep = NULL, before = 6) %>%
  # Remove a nonessential var. to ensure new col is printed
  dplyr::select(geo_value, time_value, cases, cases_7dsum) %>%
  ungroup()
#> An `epi_df` object, 4,026 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2024-01-26 17:27:32.755949
#> 
#> # A tibble: 4,026 × 4
#>    geo_value time_value cases cases_7dsum
#>  * <chr>     <date>     <dbl>       <dbl>
#>  1 ca        2020-03-01     6          NA
#>  2 ca        2020-03-02     4          NA
#>  3 ca        2020-03-03     6          NA
#>  4 ca        2020-03-04    11          NA
#>  5 ca        2020-03-05    10          NA
#>  6 ca        2020-03-06    18          NA
#>  7 ca        2020-03-07    26          81
#>  8 ca        2020-03-08    19          94
#>  9 ca        2020-03-09    23         113
#> 10 ca        2020-03-10    22         129
#> # ℹ 4,016 more rows