Generate a snapshot from an epi_archive object

Generates a snapshot in epi_df format from an epi_archive object, as of a given version. See the archive vignette for examples.

Usage

epix_as_of(x, max_version, min_time_value = -Inf)

Arguments

x

An epi_archive object

max_version

Time value specifying the max version to permit in the snapshot. That is, the snapshot will comprise the unique rows of the current archive data that represent the most up-to-date signal values, as of the specified max_version (and whose time values are at least min_time_value.)

min_time_value

Time value specifying the min time value to permit in the snapshot. Default is -Inf, which effectively means that there is no minimum considered.

Value

An epi_df object.

Details

This is simply a wrapper around the as_of() method of the epi_archive class, so if x is an epi_archive object, then:

epix_as_of(x, max_version = v)

is equivalent to:

x$as_of(max_version = v)

Examples

# warning message of data latency shown
epix_as_of(x = archive_cases_dv_subset,
           max_version = max(archive_cases_dv_subset$DT$version))
#> Warning: Getting data as of some "clobberable" version that might be hotfixed,
#> synced, or otherwise replaced later with different data using the same version
#> tag.  Thus, the snapshot that we produce here might not be reproducible later.
#> See `?epi_archive` for more info and `?epix_as_of` on how to muffle.
#> An `epi_df` object, 2,192 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2021-12-01
#> 
#> # A tibble: 2,192 × 4
#>    geo_value time_value percent_cli case_rate_7d_av
#>  * <chr>     <date>           <dbl>           <dbl>
#>  1 ca        2020-06-01        2.75            6.84
#>  2 ca        2020-06-02        2.57            6.82
#>  3 ca        2020-06-03        2.48            6.66
#>  4 ca        2020-06-04        2.41            6.98
#>  5 ca        2020-06-05        2.57            6.97
#>  6 ca        2020-06-06        2.63            6.66
#>  7 ca        2020-06-07        2.73            6.74
#>  8 ca        2020-06-08        3.04            6.67
#>  9 ca        2020-06-09        2.97            6.81
#> 10 ca        2020-06-10        2.99            7.13
#> # … with 2,182 more rows
#> # ℹ Use `print(n = ...)` to see more rows

range(archive_cases_dv_subset$DT$version) # 2020-06-02 -- 2021-12-01
#> [1] "2020-06-02" "2021-12-01"

epix_as_of(x = archive_cases_dv_subset,
           max_version = as.Date("2020-06-12"))
#> An `epi_df` object, 44 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2020-06-12
#> 
#> # A tibble: 44 × 4
#>    geo_value time_value percent_cli case_rate_7d_av
#>  * <chr>     <date>           <dbl>           <dbl>
#>  1 ca        2020-06-01        2.23            6.63
#>  2 ca        2020-06-02        2.06            6.45
#>  3 ca        2020-06-03        1.90            6.62
#>  4 ca        2020-06-04        1.79            6.64
#>  5 ca        2020-06-05        1.83            6.91
#>  6 ca        2020-06-06        1.86            6.76
#>  7 ca        2020-06-07        1.78            6.75
#>  8 ca        2020-06-08        1.90            6.90
#>  9 ca        2020-06-09       NA               7.02
#> 10 ca        2020-06-10       NA               7.36
#> # … with 34 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# When fetching a snapshot as of the latest version with update data in the
# archive, a warning is issued by default, as this update data might not yet
# be finalized (for example, if data versions are labeled with dates, these
# versions might be overwritten throughout the corresponding days with
# additional data or "hotfixes" of erroroneous data; when we build an archive
# based on database queries, the latest available update might still be
# subject to change, but previous versions should be finalized). We can
# muffle such warnings with the following pattern:
withCallingHandlers({
  epix_as_of(x = archive_cases_dv_subset,
             max_version = max(archive_cases_dv_subset$DT$version))
}, epiprocess__snapshot_as_of_clobberable_version = function(wrn) invokeRestart("muffleWarning"))
#> An `epi_df` object, 2,192 x 4 with metadata:
#> * geo_type  = state
#> * time_type = day
#> * as_of     = 2021-12-01
#> 
#> # A tibble: 2,192 × 4
#>    geo_value time_value percent_cli case_rate_7d_av
#>  * <chr>     <date>           <dbl>           <dbl>
#>  1 ca        2020-06-01        2.75            6.84
#>  2 ca        2020-06-02        2.57            6.82
#>  3 ca        2020-06-03        2.48            6.66
#>  4 ca        2020-06-04        2.41            6.98
#>  5 ca        2020-06-05        2.57            6.97
#>  6 ca        2020-06-06        2.63            6.66
#>  7 ca        2020-06-07        2.73            6.74
#>  8 ca        2020-06-08        3.04            6.67
#>  9 ca        2020-06-09        2.97            6.81
#> 10 ca        2020-06-10        2.99            7.13
#> # … with 2,182 more rows
#> # ℹ Use `print(n = ...)` to see more rows
# Since R 4.0, there is a `globalCallingHandlers` function that can be used
# to globally toggle these warnings.