Merge two epi_archive objects

Merges the underlying data tables in two epi_archive objects, allows for post-filling of NA values by last observation carried forward (LOCF), and overwrites the first data table with the merged one. See the archive vignette for examples.

epix_merge(x, y, ..., locf = TRUE, nan = NA)

Arguments

x, y
Two epi_archive objects to join together, or more specifically, whose underlying data tables are to be joined together. The data table in x will be overwritten with the joined data table. For convenience, we also allow y to be passed in directly as a data.table (need not be an epi_archive object).

...
Named arguments to pass to data.table::merge.data.table(), which is used for the join (with all default settings as in this function). For example, passing all = TRUE will perform a full join.

locf
Should LOCF be used after joining on all non-key columns? This will take the latest version of each signal value and propogate it forward to fill in gaps that appear after merging. Default is TRUE.

nan
Should NaN values be treated as NA values in the post-filling step? Default is NA, which means that they are treated as NA values; if

Value

Nothing; the data table in x is overwritten with the merged one.

Details

This is simply a wrapper around the merge() method of the epi_archive class, so if x and y are an epi_archive objects, then:

epix_merge(x, y)

is equivalent to:

x$merge(y)