Merge two epi_archive objects

Merges two epi_archives that share a common geo_value, time_value, and set of key columns. When they also share a common versions_end, using $as_of on the result should be the same as using $as_of on x and y individually, then performing a full join of the DTs on the non-version key columns (potentially consolidating multiple warnings about clobberable versions). If the versions_end values differ, the sync parameter controls what is done.

Usage

epix_merge(
  x,
  y,
  sync = c("forbid", "na", "locf", "truncate"),
  compactify = TRUE
)

Arguments

x, y

Two epi_archive objects to join together.

sync

Optional; "forbid", "na", "locf", or "truncate"; in the case that x$versions_end doesn't match y$versions_end, what do we do?: "forbid": emit an error; "na": use max(x$versions_end, y$versions_end) as the result's versions_end, but ensure that, if we request a snapshot as of a version after min(x$versions_end, y$versions_end), the observation columns from the less up-to-date archive will be all NAs (i.e., imagine there was an update immediately after its versions_end which revised all observations to be NA); "locf": use max(x$versions_end, y$versions_end) as the result's versions_end, allowing the last version of each observation to be carried forward to extrapolate unavailable versions for the less up-to-date input archive (i.e., imagining that in the less up-to-date archive's data set remained unchanged between its actual versions_end and the other archive's versions_end); or "truncate": use min(x$versions_end, y$versions_end) as the result's versions_end, and discard any rows containing update rows for later versions.

compactify

Optional; TRUE, FALSE, or NULL; should the result be compactified? See as_epi_archive for an explanation of what this means. Default here is TRUE.

Value

the resulting epi_archive

Details

This function, epix_merge, does not mutate its inputs and will not alias either archive's DT, but may alias other fields; x$merge will overwrite x with the result of the merge, reseating its DT and several other fields (making them point to different objects), but avoiding mutation of the contents of the old DT (only relevant if you have another reference to the old DT in another object).

In all cases, additional_metadata will be an empty list, and clobberable_versions_start will be set to the earliest version that could be clobbered in either input archive.

Examples

# create two example epi_archive datasets
x <- archive_cases_dv_subset$DT %>%
  dplyr::select(geo_value,time_value,version,case_rate_7d_av) %>%
  as_epi_archive(compactify=TRUE)
y <- archive_cases_dv_subset$DT %>%
  dplyr::select(geo_value,time_value,version,percent_cli) %>%
  as_epi_archive(compactify=TRUE)
# merge results stored in a third object:
xy = epix_merge(x, y)
# vs. mutating x to hold the merge result:
x$merge(y)