Slide a function over variables in an epi_archive object

Slides a given function over variables in an epi_archive object. This behaves similarly to epi_slide(), with the key exception that it is version-aware: the sliding computation at any given reference time t is performed on data that would have been available as of t. See the archive vignette for examples.

epix_slide(
  x,
  f,
  ...,
  n = 7,
  group_by,
  ref_time_values,
  time_step,
  complete = FALSE,
  new_col_name = "slide_value",
  as_list_col = FALSE,
  names_sep = "_",
  all_rows = FALSE
)

Arguments

f
Function or formula to slide over variables in x. To "slide" means to apply a function or formula over a running window of n time steps (where one time step is typically one day or one week). If a function, f must take x, a data frame with the same column names as the original object; followed by any number of named arguments; and ending with .... If a formula, f can operate directly on columns accessed via .x$var, as in ~ mean(.x$var) to compute a mean of a column var over a sliding window of n time steps.

...
Additional arguments to pass to the function or formula specified via f. Alternatively, if f is missing, then the current argument is interpreted as an expression for tidy evaluation.

n
Number of time steps to use in the running window. For example, if n = 7, and one time step is one day, then to produce a value on January 7 we apply the given function or formula to data in between January 1 and 7. Default is 7.

group_by
The variable(s) to group by before slide computation. If missing, then the keys in the underlying data table, excluding time_value and version, will be used for grouping. To omit a grouping entirely, use group_by = NULL.

ref_time_values
Time values for sliding computations, meaning, each element of this vector serves as the reference time point for one sliding window. If missing, then this will be set to all unique time values in the underlying data table, by default.

time_step
Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the time_value column. This function must take a positive integer and return an object of class lubridate::period. For example, we can use time_step = lubridate::hours in order to set the time step to be one hour (this would only be meaningful if time_value is of class POSIXct).

complete
Should the slide function be run over complete windows only? (A complete window has distance n-1 between its left and right endpoints.) Default is FALSE, which allows for partial computations.

new_col_name
String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

as_list_col
Should the new column be stored as a list column? Default is FALSE, in which case a list object returned by f would be unnested (using tidyr::unnest()), and the names of the resulting columns are given by prepending new_col_name to the names of the list elements.

names_sep
String specifying the separator to use in tidyr::unnest() when as_list_col = FALSE. Default is "_". Using NULL drops the prefix from new_col_name entirely.

all_rows
If all_rows = TRUE, then the output will have one row per combination of grouping variables and unique time values in the underlying data table. Otherwise, there will be one row in the output for each time value in x that acts as a reference time value. Default is FALSE.

Value

A tibble whose columns are: the grouping variables, time_value, containing the reference time values for the slide computation, and a column named according to the new_col_name argument, containing the slide values.

Details

Unlike epi_slide(), windows in epix_slide() are always right-aligned. Another distinction is the by argument, which used to specify the grouping upfront (whereas for an epi_df, this would be accomplished by a call to dplyr::group_by() that precedes a call to epi_df()). Apart from this, the interfaces between epix_slide() and epi_slide() are altogether similar.

The current function can be considerably slower than epi_slide(), for two reasons: (1) it must repeatedly fetch properly-versioned snapshots from the data archive (via its as_of() method), and (2) it performs a "manual" sliding of sorts, and does not benefit from the highly efficient slider package. For this reason, it should never be used in place of epi_slide(), and only used when version-aware sliding is necessary (as it its purpose).

Finally, this is simply a wrapper around the slide() method of the epi_archive class, so if x is an epi_archive object, then:

epix_slide(x, new_var = comp(old_var), n = 120)

is equivalent to:

x$slide(x, new_var = comp(old_var), n = 120)