Compute percentage change of a variable in an epi_tibble object

Computes the percentage change of a variable in an epi_tibble object. See the percentage change vignette for examples.

pct_change(
  x,
  var,
  n = 14,
  new_col_name = "pct_change",
  time_step,
  na_rm = TRUE
)

Arguments

x
The epi_tibble object.

var
The variable in x whose percentage change values are to be computed.

n
Number of time steps to use in the trailing window. For example, if n = 10, and one time step is one day, then the percentage change value computed on November 10 is 100 * (B - A) / A, where A is the sum of the variable in between November 6 and 10, and A is the sum in between November 1 and 5. Note that n must be even; default is 14.

new_col_name
String indicating the name of the new column that will contain the derivative values. Default is "slide_value"; note that setting new_col_name equal to an existing column name will overwrite this column.

time_step
Optional function used to define the meaning of one time step, which if specified, overrides the default choice based on the metadata. Read the documentation for epi_slide() for more details.

na_rm
Should missing values be removed before the computation? Default is TRUE.

Value

An epi_tibble object given by appending a new column to x, named according to the new_col_name argument, containing the percentage change values.

See also